% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_data_duty_cycle}
\alias{calculate_data_duty_cycle}
\title{Calculate weighted duty cycle}
\usage{
calculate_data_duty_cycle(act_pwr_props, tech, params)
}
\arguments{
\item{act_pwr_props}{List with proportion of time spent in low vs low-mid vs mid-high vs high output power activities}

\item{tech}{Type of technology used ("3g", "4g", or "5g")}

\item{params}{Device-specific parameter list}
}
\value{
weighted duty cycle
}
\description{
Calculate weighted duty cycle
}
