% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{calculate_data_headphone_sar}
\alias{calculate_data_headphone_sar}
\title{Calculates Data SAR for headphone positin}
\usage{
calculate_data_headphone_sar(params, tissue_params)
}
\arguments{
\item{params}{...}

\item{tissue_params}{...}
}
\description{
Calculates Data SAR for headphone positin
}
