% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_data_power_by_env}
\alias{calculate_data_power_by_env}
\title{Calculate power by environment}
\usage{
calculate_data_power_by_env(duty_cycle, urb_list, loc_props, tech, env, params)
}
\arguments{
\item{duty_cycle}{List of two elements (low and high) with corresponding duty cycle}

\item{urb_list}{Urbanicity list}

\item{loc_props}{Lost of location proportions}

\item{tech}{Type of technology used ("3g", "4g", or "5g")}

\item{env}{Environment ("home", "work", "outdoor", or "travel")}

\item{params}{Device-specific parameter list}
}
\value{
weighted duty cycle
}
\description{
Calculate power by environment
}
