% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_data_power_for_tech}
\alias{calculate_data_power_for_tech}
\title{Calculate total power by data technology}
\usage{
calculate_data_power_for_tech(tech, act_pwr_props, urb_list, loc_props, params)
}
\arguments{
\item{tech}{Type of technology used ("3g", "4g", or "5g")}

\item{act_pwr_props}{List with proportion of time spent in low vs low-mid vs mid-high vs high output power activities}

\item{urb_list}{Urbanicity list}

\item{loc_props}{List of location proportions}

\item{params}{Device-specific parameter list}
}
\description{
Calculate total power by data technology
}
