% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{calculate_mpc_data_power_by_env}
\alias{calculate_mpc_data_power_by_env}
\title{Calculate data call power by environment and tech ---------------------------}
\usage{
calculate_mpc_data_power_by_env(band, urb_list, params)
}
\arguments{
\item{band}{...}

\item{urb_list}{...}

\item{params}{...}
}
\description{
Calculate data call power by environment and tech ---------------------------
}
