% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{calculate_mpc_native_power_by_band}
\alias{calculate_mpc_native_power_by_band}
\title{Calculate native power by band ----------------------------------------------}
\usage{
calculate_mpc_native_power_by_band(band, loc_props, urb_list, params)
}
\arguments{
\item{band}{...}

\item{loc_props}{...}

\item{urb_list}{...}

\item{params}{...}
}
\description{
Calculate native power by band ----------------------------------------------
}
