% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{calculate_mpc_native_power_by_env}
\alias{calculate_mpc_native_power_by_env}
\title{Calculate native power by environment and band ------------------------------}
\usage{
calculate_mpc_native_power_by_env(band, urb_list, params)
}
\arguments{
\item{band}{...}

\item{urb_list}{...}

\item{params}{...}
}
\description{
Calculate native power by environment and band ------------------------------
}
