% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{calculate_native_headphone_sar}
\alias{calculate_native_headphone_sar}
\title{Calculates Native SAR for headphone position}
\usage{
calculate_native_headphone_sar(params, tissue_params)
}
\arguments{
\item{params}{...}

\item{tissue_params}{...}
}
\description{
Calculates Native SAR for headphone position
}
