% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_power_by_env}
\alias{calculate_power_by_env}
\title{Calculate power by environment}
\usage{
calculate_power_by_env(duty_cycle, urb_list, loc_props, tech, env, params)
}
\arguments{
\item{duty_cycle}{...}

\item{urb_list}{...}

\item{loc_props}{...}

\item{tech}{...}

\item{env}{...}

\item{params}{...}
}
\value{
weighted duty cycle
}
\description{
Calculate power by environment
}
