% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_total_power_for_tech}
\alias{calculate_total_power_for_tech}
\title{Calculate total power by data technology}
\usage{
calculate_total_power_for_tech(
  tech,
  act_pwr_props,
  urb_list,
  loc_props,
  params
)
}
\arguments{
\item{tech}{...}

\item{act_pwr_props}{...}

\item{urb_list}{...}

\item{loc_props}{...}

\item{params}{...}
}
\description{
Calculate total power by data technology
}
