% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_total_wifi_power}
\alias{calculate_total_wifi_power}
\title{Calculate total wifi output power}
\usage{
calculate_total_wifi_power(act_pwr_props, params)
}
\arguments{
\item{act_pwr_props}{List with proportion of time spent in low vs low-mid vs mid-high vs high output power activities}

\item{params}{device-specific parameters}
}
\description{
Calculate total wifi output power
}
