% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_weighted_duty_cycle}
\alias{calculate_weighted_duty_cycle}
\title{Calculate weighted duty cycle}
\usage{
calculate_weighted_duty_cycle(act_pwr_props, tech, params)
}
\arguments{
\item{act_pwr_props}{...}

\item{tech}{"3g", "4g", or "5g"}

\item{params}{...}
}
\value{
weighted duty cycle
}
\description{
Calculate weighted duty cycle
}
