% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{calculate_wifi_band_headphone_sar}
\alias{calculate_wifi_band_headphone_sar}
\title{Calculate WiFi SAR from phone while used with headphones for a specific frequency band and phone position}
\usage{
calculate_wifi_band_headphone_sar(band, params, tissue_params)
}
\arguments{
\item{band}{frequency band ("2" for 2.4GHz, "5" for 5.0GHz)}

\item{params}{...}

\item{tissue_params}{...}
}
\description{
Calculate WiFi SAR from phone while used with headphones for a specific frequency band and phone position
}
