% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{calculate_wifi_band_total_sar}
\alias{calculate_wifi_band_total_sar}
\title{Calculate the SAR from WiFi for a specific frequency band (2.4 GHz or 5.0 GHz)}
\usage{
calculate_wifi_band_total_sar(
  band,
  ear_prop,
  headp_prop,
  speaker_prop,
  params,
  tissue_params
)
}
\arguments{
\item{band}{frequency band ("2" for 2.4GHz, "5" for 5.0GHz)}

\item{ear_prop}{...}

\item{headp_prop}{...}

\item{speaker_prop}{...}

\item{params}{...}

\item{tissue_params}{...}
}
\description{
Calculate the SAR from WiFi for a specific frequency band (2.4 GHz or 5.0 GHz)
}
