% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{calculate_wifi_duty_cycle}
\alias{calculate_wifi_duty_cycle}
\title{Calculate WiFi duty cycle}
\usage{
calculate_wifi_duty_cycle(act_pwr_props, band, params)
}
\arguments{
\item{act_pwr_props}{List with proportion of time spent in low vs low-mid vs mid-high vs high output power activities}

\item{band}{Frequency band ("2" for 2.4 GHz, "5" for 5.0 GHz)}

\item{params}{device-specific parameters}
}
\description{
Calculate WiFi duty cycle
}
