% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_wifi_exposure_duration}
\alias{calculate_wifi_exposure_duration}
\title{Calculate WiFi exposure duration}
\usage{
calculate_wifi_exposure_duration(
  travel_time,
  wifi_prop_travel,
  home_prop,
  work_prop
)
}
\arguments{
\item{travel_time}{time spent commutng in car/train/bus per day in s}

\item{wifi_prop_travel}{proportion of time connected to WiFi (vs mobile data) during commute}

\item{home_prop}{proportion of time per day spent at home}

\item{work_prop}{proportion of time per day spent at school/work}
}
\value{
exposure duration in s
}
\description{
Calculate WiFi exposure duration
}
