% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_country}
\alias{check_country}
\title{Check input values - country}
\usage{
check_country(country)
}
\arguments{
\item{country}{Country}
}
\value{
re-coded urbanicity (binary variables)
}
\description{
Check input values - country
}
