% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_duration}
\alias{check_duration}
\title{Check input duration(s)}
\usage{
check_duration(duration)
}
\arguments{
\item{duration}{single duration (numeric) or vector of durations}
}
\value{
Returns invisible TRUE if all checks pass. Side effects of stopping
or showing warning message if any check fails
}
\description{
Checks supplied input duration or vector of durations; needed to stop
calculation or raise warning in case of irregular input values.
}
