% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_input_param_list}
\alias{check_input_param_list}
\title{Check input parameter list}
\usage{
check_input_param_list(params)
}
\arguments{
\item{params}{parameter list in YAML format}
}
\value{
TRUE if parameter list is valid, FALSE if parameter list is invalid
}
\description{
Check input parameter list
}
