% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_proportions}
\alias{check_proportions}
\title{Check input proportion(s)}
\usage{
check_proportions(proportions)
}
\arguments{
\item{proportions}{Single proportion value (between 0 and 1 ) or vector
with multiple proportions to check}
}
\value{
Returns invisible TRUE if all checks pass. Side effects of stopping
or showing warning message if any check fails.
}
\description{
Checks supplied input proportion or vector of proportions; needed to stop
calculation or raise warning in case of irregular input values.
}
