% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_urbanicity}
\alias{check_urbanicity}
\title{Check input value s- urbanicity}
\usage{
check_urbanicity(urbanicity)
}
\arguments{
\item{urbanicity}{urbanicity}
}
\description{
Check input value s- urbanicity
}
