% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{combine_mpd_source_power}
\alias{combine_mpd_source_power}
\title{Combine source power}
\usage{
combine_mpd_source_power(wifi_prop, wifi_pwr, data_pwr)
}
\arguments{
\item{wifi_prop}{Proportion of time connected to WiFi (vs data)}

\item{wifi_pwr}{Output power from WiFi in mW}

\item{data_pwr}{Output power from data in mW}
}
\value{
combined power from WiFi and data in mW
}
\description{
Combines the output power from WiFi and Data by scaling it with their use
proportions.
}
