% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordless.R
\name{dect_pwr}
\alias{dect_pwr}
\title{Calculate Output Power of DECT/Cordless Phone}
\usage{
dect_pwr(params = load_params())
}
\arguments{
\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
DECT/cordless phone output power in mW
}
\description{
Calculates the output power of DECT/cordless phone during a call.
}
