% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordless.R
\name{dect_sar}
\alias{dect_sar}
\title{Calculate nSAR from Cordless Calls}
\usage{
dect_sar(tissue, ear_prop, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate the nSAR (default: "brain" or "body")}

\item{ear_prop}{Proportion of time cordless phone is held against ear
during calls.}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific nSAR from DECT/cordless phone calls in W/kg/W
}
\description{
Calculates nSAR (normalized specific absorption rate) from making calls with
cordless phone (also known as DECT phone).
}
\examples{
dect_sar(
  tissue      = "brain",
  ear_prop    = 0.9)

}
