% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_default_variables}
\alias{download_default_variables}
\title{Download default variable list}
\usage{
download_default_variables(file = NULL)
}
\arguments{
\item{file}{location where downloaded YAML file should be saved, must be a YAML filename}
}
\description{
Download default variable list
}
