% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data for testing}
\format{
A data frame
\describe{
\item{sample_id}{test}
\item{use_5g}{test}
\item{travel_time}{test}
\item{urbanicity}{test}
\item{country}{test}
\item{headp_ear_num}{test}
\item{mpc_duration}{test}
\item{mpc_ear_prop}{test}
\item{mpc_headp_prop}{test}
\item{mpd_wifi_prop_home}{test}
\item{mpd_wifi_prop_travel}{test}
\item{mpd_wifi_prop_work}{test}
\item{mpd_dur_low}{test}
\item{mpd_dur_lowtomed}{test}
\item{mpd_dur_medtohigh}{test}
\item{mpd_dur_high}{test}
\item{dect_duration}{test}
\item{dect_ear_prop}{test}
\item{lptp_dur_low}{test}
\item{lptp_dur_lowtomed}{test}
\item{lptp_dur_medtohigh}{test}
\item{lptp_dur_high}{test}
\item{tblt_dur_low}{test}
\item{tblt_dur_lowtomed}{test}
\item{tblt_dur_medtohigh}{test}
\item{tblt_dur_high}{test}
\item{hotspot_duration}{test}
\item{smartwatch_duration}{test}
\item{tracker_duration}{test}
\item{vr_duration}{test}
\item{headphone_duration}{test}
\item{gaming_duration}{test}
}
}
\source{
Generated for package testing.
}
\usage{
example_data
}
\description{
A dataset containing example user data (made up)
}
\keyword{datasets}
