% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{farfield_dose}
\alias{farfield_dose}
\title{Calculate RF-EMF Dose from Far-field Sources}
\usage{
farfield_dose(tissue, country, urbanicity, travel_time, params = load_params())
}
\arguments{
\item{country}{Country of residence (Austria = "AT", Belgium = "BE", France = FR,
Hungary = "HU", Italy = "IT", Netherlands = "NL", Poland = "PL", Spain = "ES",
Switzerland = "CH", United Kingdom = "UK", elsewhere = "Other")}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from far-field sources in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from far-field sources
}
\details{
The RF-EMF dose from far-field sources is calculated as:

\deqn{Dose_{farfield} = mSAR_{farfield}*86400}

Where:
\itemize{
\item \eqn{Dose_{farfield}} is the dose from tablet use
\item \eqn{mSAR_{farfield}} is the momentary SAR value in mJ/kg
}
}
\examples{
farfield_dose(
  tissue      = "body",
  country     = "CH",
  urbanicity  = "suburban",
  travel_time = 1800)

}
\seealso{
\code{\link[=farfield_msar]{farfield_msar()}}
}
