% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{farfield_msar}
\alias{farfield_msar}
\title{Calculate mSAR of Far-field Sources}
\usage{
farfield_msar(tissue, country, urbanicity, travel_time, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{country}{Country of residence (Austria = "AT", Belgium = "BE", France = FR,
Hungary = "HU", Italy = "IT", Netherlands = "NL", Poland = "PL", Spain = "ES",
Switzerland = "CH", United Kingdom = "UK", elsewhere = "Other")}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
mSAR in W/kg/W/m**2
}
\description{
Calculates the mSAR of far-field sources
}
\details{
The output power depends on the country, urbanicity, and time spent commuting.
}
\examples{
farfield_msar(
  tissue      = "body",
  country     = "CH",
  urbanicity  = "suburban",
  travel_time = 1800)

}
\seealso{
\code{\link[=farfield_pwr]{farfield_pwr()}}, \code{\link[=farfield_sar]{farfield_sar()}}
}
