% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{farfield_pwr}
\alias{farfield_pwr}
\title{Calculate Output Power of Far-field Sources}
\usage{
farfield_pwr(country, urbanicity, travel_time, params = load_params())
}
\arguments{
\item{country}{Country of residence (Austria = "AT", Belgium = "BE", France = FR,
Hungary = "HU", Italy = "IT", Netherlands = "NL", Poland = "PL", Spain = "ES",
Switzerland = "CH", United Kingdom = "UK", elsewhere = "Other")}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power in mW/m**2
}
\description{
Calculates the output power of far-field sources
}
\details{
The output power depends on the country, urbanicity, and time spent commuting.
}
\examples{
farfield_pwr(
country     = "CH",
urbanicity  = "suburban",
travel_time = 1800)

}
