% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{flat_params_to_nested}
\alias{flat_params_to_nested}
\title{Convert flat parameter table to nested parameter YAML}
\usage{
flat_params_to_nested(flat)
}
\arguments{
\item{flat}{Flat parameter data frame with "path" and "value" columns}
}
\value{
nested parameter list
}
\description{
Convert flat parameter table to nested parameter YAML
}
