% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaming.R
\name{gaming_dose}
\alias{gaming_dose}
\title{Calculate RF-EMF Dose from online gaming (console)}
\usage{
gaming_dose(tissue, duration_gaming, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_gaming}{duration of gaming with portable console in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from gaming in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from online gaming
}
\details{
The RF-EMF dose from online gaming is calculated as:

\deqn{Dose_{gaming} = mSAR_{gaming} * duration_{gaming}}

Where:
\itemize{
\item \eqn{mSAR_{gaming}} is the mSAR of gaming
\item \eqn{Dduration_{gaming}} is the duration of gaming with portable console in s/day
}
}
\examples{
gaming_dose(
  tissue          = "body",
  duration_gaming = 600)

}
