% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordless.R
\name{get_cordless_pwr}
\alias{get_cordless_pwr}
\title{Calculate Aggregated Power of Cordless Phone}
\usage{
get_cordless_pwr(params)
}
\arguments{
\item{params}{descr}
}
\value{
descr
}
\description{
Calculate Aggregated Power of Cordless Phone
}
