% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{get_farfield_dose}
\alias{get_farfield_dose}
\title{Calculate Dose from far-field exposure}
\usage{
get_farfield_dose(urbanicity, travel_time, params)
}
\arguments{
\item{urbanicity}{Urbanicity of home environment}

\item{travel_time}{Time per day spent commuting (s)}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculates RF-EMF dose from far-field exposure
}
\details{
The far-field RF-EMF dose is calculated as:
\deqn{Dose_{farfield} = 86400s \times Power_{farfield} \times SAR_{farfield}}

Where:
\itemize{
\item \eqn{Power_{farfield}} is the output power of farfield sources
\item \eqn{SAR_{mobiledata}} is the specific absorption rate in W/kg/W
}
}
\seealso{
\code{\link[=get_farfield_pwr]{get_farfield_pwr()}},\code{\link[=get_farfield_sar]{get_farfield_sar()}}
}
