% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{get_farfield_pwr_by_country}
\alias{get_farfield_pwr_by_country}
\title{Get far-field output power by country}
\usage{
get_farfield_pwr_by_country(country, base, params)
}
\arguments{
\item{country}{country}

\item{base}{Base name}

\item{params}{Parameter list}
}
\description{
Get far-field output power by country
}
