% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_gaming_dose}
\alias{get_gaming_dose}
\title{Calculate brain and body dose from gaming with a portable console}
\usage{
get_gaming_dose(duration_gaming, params = NULL)
}
\arguments{
\item{duration_gaming}{Daily gaming duration with portable console in seconds}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from smart home
}
\description{
Calculate brain and body dose from gaming with a portable console
}
