% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_headp_phone_dose}
\alias{get_headp_phone_dose}
\title{Calculate contribution from mobile phone while using headp}
\usage{
get_headp_phone_dose(duration_headphones, params, tissue_params)
}
\arguments{
\item{duration_headphones}{Daily duration of headphone use in seconds}

\item{params}{Parameter list}

\item{tissue_params}{tissue-specific parameter list}
}
\description{
Calculate contribution from mobile phone while using headp
}
