% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_headp_pwr}
\alias{get_headp_pwr}
\title{Calculate power from bluetooth headphones (headphone contribution only)}
\usage{
get_headp_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
output power
}
\description{
Calculate power from bluetooth headphones (headphone contribution only)
}
