% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_hotspot_dose}
\alias{get_hotspot_dose}
\title{Calculate brain and body dose from using mobile phone as hotspot}
\usage{
get_hotspot_dose(duration_hotspot, params)
}
\arguments{
\item{duration_hotspot}{Daily duration of using phone as hotspot in seconds}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from hotspot use
}
\description{
Calculate brain and body dose from using mobile phone as hotspot
}
