% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_hotspot_pwr}
\alias{get_hotspot_pwr}
\title{Calculate output power from using mobile phone as hotspot}
\usage{
get_hotspot_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Aggregated output power of mobile phone hotspot in mW
}
\description{
Calculate output power from using mobile phone as hotspot
}
