% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_hotspot_sar}
\alias{get_hotspot_sar}
\title{Calculate SAR from hotspot}
\usage{
get_hotspot_sar(params, tissue_params)
}
\arguments{
\item{params}{Parameter list}

\item{tissue_params}{Tissue params}
}
\value{
tissue-specific SAR from hotspot
}
\description{
Calculate SAR from hotspot
}
