% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptop.R
\name{get_laptop_sar}
\alias{get_laptop_sar}
\title{Calculate Laptop Aggregated SAR}
\usage{
get_laptop_sar(params, tissue_params)
}
\arguments{
\item{params}{descr}

\item{tissue_params}{descr}
}
\value{
Aggregated tissue SAR from laptop use (W/kg/W)
}
\description{
Calculate Laptop Aggregated SAR
}
