% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_bt_headp_dose}
\alias{get_mobilecall_bt_headp_dose}
\title{Calculate RF-EMF dose from mobile calling using Bluetooth headphones (headphone contribution)}
\usage{
get_mobilecall_bt_headp_dose(
  duration,
  headp_prop,
  headp_ear_num,
  params,
  tissue_params
)
}
\arguments{
\item{duration}{Duration of mobile phone call in seconds per day}

\item{headp_prop}{Proportion of time Bluetooth headphones are used during mobile call}

\item{headp_ear_num}{If particpant uses 1 or 2 Bluetooth headphones during mobile call}

\item{params}{Parameter list}

\item{tissue_params}{Tissue parameter list}
}
\value{
Dose from mobile calling using Bluetooth headphones in mJ/kg/day (headphone contribution)
}
\description{
Calculates the RF-EMF dose from mobile calling using Bluetooth headphones (headphone contribution).
}
\details{
The dose is calculated as:
\deqn{Dose = Duration \times Power \times SAR \times headp_prop \times headp_ear_num}

Where:
\itemize{
\item \eqn{Duration} is the duration of mobile phone call in seconds per day
\item \eqn{SAR} is the specific absorption rate from Bluetooth in W/kg/W
\item \eqn{Power} is the Bluetooth output power of the mobile phone during mobile calling mW
\item \eqn{headp_prop} is the proportion of time Bluetooth headphones are used during mobile calls
\item \eqn{headp_ear_num} is the number of Bluetooth headphones (1 or 2) used during mobile calls
}
}
\seealso{
\code{\link[=get_mobilecall_bt_phone_dose]{get_mobilecall_bt_phone_dose()}}, \code{\link[=get_mobilecall_bt_headp_pwr]{get_mobilecall_bt_headp_pwr()}}, \code{\link[=get_mobilecall_bt_headp_sar]{get_mobilecall_bt_headp_sar()}}
}
