% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_bt_phone_dose}
\alias{get_mobilecall_bt_phone_dose}
\title{Mobilecall bluetooth phone dose}
\usage{
get_mobilecall_bt_phone_dose(duration, params, tissue_params)
}
\arguments{
\item{duration}{duration}

\item{params}{parameter list}

\item{tissue_params}{tissue parameters}
}
\description{
Mobilecall bluetooth phone dose
}
