% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_bt_phone_pwr}
\alias{get_mobilecall_bt_phone_pwr}
\title{Mobilecall bluetooth phone power}
\usage{
get_mobilecall_bt_phone_pwr(params)
}
\arguments{
\item{params}{parameter list}
}
\description{
Mobilecall bluetooth phone power
}
