% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_bt_phone_sar}
\alias{get_mobilecall_bt_phone_sar}
\title{Calculate SAR from mobile calling using Bluetooth headphones (phone contribution)}
\usage{
get_mobilecall_bt_phone_sar(params, tissue_params)
}
\arguments{
\item{params}{Parameter list}

\item{tissue_params}{Tissue parameter list}
}
\value{
SAR from mobile calling using Bluetooth headphones in W/kg/W (phone contribution)
}
\description{
Calculates the specific absorption rate from mobile calling using Bluetooth headphones (phone contribution).
}
