% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_data_pwr}
\alias{get_mobilecall_phone_data_pwr}
\title{Mobilecall Data Power}
\usage{
get_mobilecall_phone_data_pwr(urbanicity, params)
}
\arguments{
\item{urbanicity}{Urbanicity}

\item{params}{Parameter list}
}
\value{
mobilecall data power
}
\description{
TODO: adapt calculations to urbanicity levels
}
