% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_native_sar}
\alias{get_mobilecall_phone_native_sar}
\title{Calculate SAR from native mobile calling (no Bluetooth)}
\usage{
get_mobilecall_phone_native_sar(
  ear_prop,
  headp_prop,
  speaker_prop,
  params,
  tissue_params
)
}
\arguments{
\item{ear_prop}{Proportion of time mobile phone is held against ear during call}

\item{headp_prop}{Proportion of time Bluetooth headphones are used during mobile call while phone is NOT held against ear}

\item{speaker_prop}{Proportion of time speaker mode is used during mobile call while phone is NOT held against ear}

\item{params}{Parameter list}

\item{tissue_params}{Tissue parameter list}
}
\value{
SAR value from mobile phone during native calls in W/kg/W
}
\description{
Calculates the tissue-specific absorption rate (SAR) for calling with a mobile phone (no Bluetooth) using native network.
}
\details{
The SAR from native mobile calling is calculated as:
\deqn{SAR_{Native} = Prop_{Ear} \times SAR_{Ear} + Prop_{Speaker} \times SAR_{Speaker} + Prop_{Headphones} \times SAR_{Headphones}}

Where:
\itemize{
\item \eqn{Prop_{Ear}}, \eqn{Prop_{Speaker}}, \eqn{Prop_{Headphones}} are the proportion of time phone is held against ear, used in speaker mode, and used with headphones, respectively
\item \eqn{SAR_{Ear}}, \eqn{SAR_{Speaker}}, \eqn{SAR_{Headphones}} are the SAR values from holding phone to ear, using it in speaker mode, and using it with headphones, respectively
}
}
