% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_wifi_sar}
\alias{get_mobilecall_phone_wifi_sar}
\title{Calculate SAR from WiFi mobile calling (no Bluetooth)}
\usage{
get_mobilecall_phone_wifi_sar(
  ear_prop,
  headp_prop,
  speaker_prop,
  params,
  tissue_params
)
}
\arguments{
\item{ear_prop}{Proportion of time mobile phone is held against ear during call}

\item{headp_prop}{Proportion of time Bluetooth headphones are used during mobile call while phone is NOT held against ear}

\item{speaker_prop}{Proportion of time speaker mode is used during mobile call while phone is NOT held against ear}

\item{params}{Parameter list}

\item{tissue_params}{Tissue parameter list}
}
\value{
SAR value from mobile phone during WiFi calls in W/kg/W
}
\description{
Calculates the tissue-specific absorption rate (SAR) for calling with a mobile phone (no Bluetooth) using WiFi.
}
\details{
The SAR from mobile calling using WiFi is calculated as:
\deqn{SAR_{WiFi} = Prop_{2.4GHz} \times SAR_{2.4GHz} + Prop_{5.0GHz} \times SAR_{5.0GHz}}

Where:
\itemize{
\item \eqn{Prop_{2.4GHz}}, \eqn{Prop_{5.0GHz}} are the proportion of 2.4GHz and 5.0GHz WiFi, respectively
\item \eqn{SAR_{2.4GHz}}, \eqn{SAR_{5.0GHz}} are the SAR values from 2.4GHz and 5.0GHz WiFi calls from the mobile phone (no Bluetooth), respectively
}
}
