% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_pwr}
\alias{get_mobilecall_pwr}
\title{Calculate mobile call aggregated power}
\usage{
get_mobilecall_pwr(headp_prop, params)
}
\arguments{
\item{headp_prop}{Proportion of time the phone is used with bluetooth headphones during call
(needed for calculating bluetooth contribution)}

\item{params}{Mobile call related parameter list}
}
\value{
mobile call aggregated power from phone and bluetooth headphones
}
\description{
Note that the parameters ear_prop, speaker_prop and headphone_prop must add up
to one!
}
