% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_pwr_headp_bt}
\alias{get_mobilecall_pwr_headp_bt}
\title{Calculate mobile call aggregated power from bluetooth headphones}
\usage{
get_mobilecall_pwr_headp_bt(headp_prop, params)
}
\arguments{
\item{headp_prop}{Proportion of time the phone is used with bluetooth headphones during call
(needed for calculating bluetooth contribution)}

\item{params}{description}
}
\value{
aggregated output power of bluetooth headphones during mobile call
scaled with headphone use proportion, multiplied by two (headphone in two ears)
}
\description{
Calculate mobile call aggregated power from bluetooth headphones
}
