% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_pwr_phone}
\alias{get_mobilecall_pwr_phone}
\title{Calculate mobile call aggregated power from phone itself}
\usage{
get_mobilecall_pwr_phone(params)
}
\arguments{
\item{params}{Mobile call related parameter list}
}
\value{
aggregated power from mobile call from phone (no bluetooth)
}
\description{
We calculate the total output power of the mobile phone itself (no bluetooth)
by multiplying the proportion of the time a technology is used times its
output power.
\deqn{AggrPwr = test}
At the moment, the output power of each technology is a constant, but this
will be refined to take user-specific parameters into account in the future.
}
