% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_pwr_phone_bt}
\alias{get_mobilecall_pwr_phone_bt}
\title{Calculate mobile call aggregated power from phone bluetooth}
\usage{
get_mobilecall_pwr_phone_bt(headp_prop, params)
}
\arguments{
\item{headp_prop}{Proportion of time the phone is used with bluetooth headphones during call
(needed for calculating bluetooth contribution)}

\item{params}{mobile call related parameter list}
}
\value{
aggregated output power of phone transmitting via bluetooth
scaled with headphone use proportion
}
\description{
Calculate mobile call aggregated power from phone bluetooth
}
