% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_sar}
\alias{get_mobilecall_sar}
\title{Calculate mobile call SAR}
\usage{
get_mobilecall_sar(ear_prop, speaker_prop, headp_prop, params, tissue_params)
}
\arguments{
\item{ear_prop}{Proportion of time the phone is held against the ear during call}

\item{speaker_prop}{Proportion of time the phone is used in speaker mode}

\item{headp_prop}{Proportion of time the phone is used with bluetooth headphones}

\item{params}{mobile call specific parameter list}

\item{tissue_params}{tissue-specific parameter list (SAR values)}
}
\value{
aggregated tissue SAR from mobile call, including bluetooth headphones
}
\description{
Note that ear_prop, speaker_prop and headp_prop should add up to one!
}
