% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_sar_data}
\alias{get_mobilecall_sar_data}
\title{Calculate SAR from mobile data phone calling (no bluetooth)}
\usage{
get_mobilecall_sar_data()
}
\description{
Calculate SAR from mobile data phone calling (no bluetooth)
}
