% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_sar_headp_bt}
\alias{get_mobilecall_sar_headp_bt}
\title{Calculate tissue SAR from bluetooth headphones during mobile call}
\usage{
get_mobilecall_sar_headp_bt(params, tissue_params)
}
\arguments{
\item{params}{description}

\item{tissue_params}{description}
}
\value{
tissue SAR from bluetooth emitted by bluetooth headphones
}
\description{
Calculate tissue SAR from bluetooth headphones during mobile call
}
