% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_sar_native}
\alias{get_mobilecall_sar_native}
\title{Calculate SAR from native phone calling (no bluetooth)}
\usage{
get_mobilecall_sar_native()
}
\description{
Calculate SAR from native phone calling (no bluetooth)
}
