% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_sar_phone}
\alias{get_mobilecall_sar_phone}
\title{Calculate mobile call SAR from mobile phone itself}
\usage{
get_mobilecall_sar_phone(
  ear_prop,
  speaker_prop,
  headp_prop,
  params,
  tissue_params
)
}
\arguments{
\item{ear_prop}{Proportion of time the phone is held against the ear during call}

\item{speaker_prop}{Proportion of time the phone is used in speaker mode}

\item{headp_prop}{Proportion of time the phone is used with bluetooth headphones}

\item{params}{description}

\item{tissue_params}{descr}
}
\value{
description
}
\description{
This is done by adding up the contributions of native, data and wifi technologies.
The native contribution is calculated like this:
\deqn{AA_5 \times \left( (O_5 \times U_5) + (P_5 \times V_5) + (Z_5 \times T_5) \right)}
}
\details{
Note that ear_prop, speaker_prop and headp_prop should add up to one!
}
