% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_sar_phone_bt}
\alias{get_mobilecall_sar_phone_bt}
\title{Calculate tissue SAR from phone during mobile call transmitting to bluetooth headphones}
\usage{
get_mobilecall_sar_phone_bt(params, tissue_params)
}
\arguments{
\item{params}{description}

\item{tissue_params}{description}
}
\value{
tissue SAR from bluetooth emitted by phone
}
\description{
Calculate tissue SAR from phone during mobile call transmitting to bluetooth headphones
}
