% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_sar_wifi}
\alias{get_mobilecall_sar_wifi}
\title{Calculate SAR from WiFi phone calling (no bluetooth)}
\usage{
get_mobilecall_sar_wifi()
}
\description{
Calculate SAR from WiFi phone calling (no bluetooth)
}
