% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_high_pwr}
\alias{get_mobiledata_high_pwr}
\title{Calculate mobile data output power from high power activities}
\usage{
get_mobiledata_high_pwr(mpd_high_dt_dur, params)
}
\arguments{
\item{mpd_high_dt_dur}{daily duration of high output power activities (s)}

\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Examples for high output power activities: ...
}
