% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_low_pwr}
\alias{get_mobiledata_low_pwr}
\title{Calculate mobile data output power from low power activities}
\usage{
get_mobiledata_low_pwr(mpd_low_dt_dur, params)
}
\arguments{
\item{mpd_low_dt_dur}{daily duration of low output power activities (s)}

\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Examples for low output power activities: ...
}
