% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_lowmed_pwr}
\alias{get_mobiledata_lowmed_pwr}
\title{Calculate mobile data output power from low-medium power activities}
\usage{
get_mobiledata_lowmed_pwr(mpd_lowmed_dt_dur, params)
}
\arguments{
\item{mpd_lowmed_dt_dur}{daily duration of low-medium output power activities (s)}

\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Examples for low-medium output power activities: ...
}
