% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_medhigh_pwr}
\alias{get_mobiledata_medhigh_pwr}
\title{Calculate mobile data output power from medium-high power activities}
\usage{
get_mobiledata_medhigh_pwr(mpd_medhigh_dt_dur, params)
}
\arguments{
\item{mpd_medhigh_dt_dur}{daily duration of medium-high output power activities (s)}

\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Examples for medium-high output power activities: ...
}
