% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_pwr}
\alias{get_mobiledata_pwr}
\title{Calculate mobile data aggregated power}
\usage{
get_mobiledata_pwr(wifi_prop, use_5g, data_prop, high_pwr_prop, params)
}
\arguments{
\item{wifi_prop}{wifi proportion}

\item{data_prop}{data proportion}

\item{high_pwr_prop}{Proportion of time spent on high data transfer activities}

\item{params}{parameter list}
}
\value{
aggregated power
}
\description{
Calculate mobile data aggregated power
}
