% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_sar}
\alias{get_mobiledata_sar}
\title{Calculate SAR from data use on mobile phone (mobile data and WiFi)}
\usage{
get_mobiledata_sar(wifi_prop, params, tissue_params)
}
\arguments{
\item{wifi_prop}{Proportion of time WiFi connection is used for data transfer (vs mobile data)}

\item{params}{Parameter list}

\item{tissue_params}{Tissue-specific parameters (SAR-values)}
}
\description{
Calculates specific absorption rate (SAR) from data use on mobile phone (mobile data and WiFi)
}
\details{
The SAR is calculated as follows:
\deqn{SAR = Prop_{WiFi}\times SAR_{WiFi} + Prop_{Data}\times SAR_{data}}

Where:
\eqn{Prop_{WiFi}}, \eqn{Prop_{Data}} are the WiFi and data use proportions
\eqn{SAR_{WiFi}}, \eqn{SAR_{Data}} are the SAR values from WiFi and from data
}
