% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_wifi_pwr}
\alias{get_mobiledata_wifi_pwr}
\title{Calculate mobile data output power from WiFi}
\usage{
get_mobiledata_wifi_pwr(params)
}
\arguments{
\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Calculate mobile data output power from WiFi
}
