% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mpd_data_3g_pwr}
\alias{get_mpd_data_3g_pwr}
\title{Calculate mpd power from 3G}
\usage{
get_mpd_data_3g_pwr(urbanicity, loc_props, act_pwr_props, params)
}
\arguments{
\item{urbanicity}{urbanicity}

\item{loc_props}{loc props}

\item{act_pwr_props}{...}

\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Calculate mpd power from 3G
}
