% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mpd_data_5g_pwr}
\alias{get_mpd_data_5g_pwr}
\title{Calculate mpd power from 5G}
\usage{
get_mpd_data_5g_pwr(urbanicty, loc_props, act_pwr_props, params)
}
\arguments{
\item{loc_props}{loc props}

\item{act_pwr_props}{...}

\item{params}{device-specific parameters}

\item{urbanicity}{urbanicity}
}
\value{
pwr
}
\description{
Calculate mpd power from 5G
}
