% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mpd_wifi_5_pwr}
\alias{get_mpd_wifi_5_pwr}
\title{Calculate wifi power from 5.0ghz}
\usage{
get_mpd_wifi_5_pwr(params)
}
\arguments{
\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Calculate wifi power from 5.0ghz
}
