% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mpd_wifi_pwr}
\alias{get_mpd_wifi_pwr}
\title{Calculate mpd power from wifi}
\usage{
get_mpd_wifi_pwr(params)
}
\arguments{
\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Calculate mpd power from wifi
}
