% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smarthome_dose}
\alias{get_smarthome_dose}
\title{Calculate brain and body dose from smart home}
\usage{
get_smarthome_dose(duration_smarthome, params = NULL)
}
\arguments{
\item{duration_smarthome}{Daily duration of being in smart home}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from smart home
}
\description{
Calculate brain and body dose from smart home
}
