% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smarthome_pwr}
\alias{get_smarthome_pwr}
\title{Calculate output power from smart home}
\usage{
get_smarthome_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Aggregated output power smart home in mW
}
\description{
Calculate output power from smart home
}
