% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smarthome_sar}
\alias{get_smarthome_sar}
\title{Calculate SAR from smart home}
\usage{
get_smarthome_sar(params, tissue_params)
}
\arguments{
\item{params}{Parameter list}

\item{tissue_params}{Tissue params}
}
\value{
tissue-specific SAR from smart home
}
\description{
Calculate SAR from smart home
}
