% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smartwatch_dose}
\alias{get_smartwatch_dose}
\title{Calculate brain and body dose from using smart watch connected to mobile phone}
\usage{
get_smartwatch_dose(duration_smartwatch, params)
}
\arguments{
\item{duration_smartwatch}{Daily duration of using smart watch in seconds}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from smart watch use
}
\description{
Important: exposure comes from both mobile phone and from smart watch.
}
