% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smartwatch_phone_dose}
\alias{get_smartwatch_phone_dose}
\title{Calculate contribution from mobile phone while using smart watch}
\usage{
get_smartwatch_phone_dose(duration_smartwatch, params, tissue_params)
}
\arguments{
\item{duration_smartwatch}{Daily duration of smartwatch use in seconds}

\item{params}{Parameter list}

\item{tissue_params}{tissue-specific parameter list}
}
\description{
Calculate contribution from mobile phone while using smart watch
}
