% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smartwatch_phone_pwr}
\alias{get_smartwatch_phone_pwr}
\title{Calculate power from smart watch (phone contribution only)}
\usage{
get_smartwatch_phone_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
output power
}
\description{
Calculate power from smart watch (phone contribution only)
}
