% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smartwatch_pwr}
\alias{get_smartwatch_pwr}
\title{Calculate power from smart watch (watch contribution only)}
\usage{
get_smartwatch_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
output power
}
\description{
Calculate power from smart watch (watch contribution only)
}
