% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{get_tablet_dose}
\alias{get_tablet_dose}
\title{Calculate Dose from Tablet Use}
\usage{
get_tablet_dose(dur_low, dur_lowtomed, dur_medtohigh, dur_high, params = NULL)
}
\arguments{
\item{dur_low}{tba}

\item{dur_lowtomed}{tba}

\item{dur_medtohigh}{tba}

\item{dur_high}{tba}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from Tablet Use
}
