% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{get_tablet_sar}
\alias{get_tablet_sar}
\title{Calculate Tablet Aggregated SAR}
\usage{
get_tablet_sar(params, tissue_params)
}
\arguments{
\item{params}{descr}

\item{tissue_params}{descr}
}
\value{
Aggregated tablet SAR
}
\description{
Calculate Tablet Aggregated SAR
}
