% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_tracker_pwr}
\alias{get_tracker_pwr}
\title{Calculate power from tracker (tracker contribution only)}
\usage{
get_tracker_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
output power
}
\description{
Calculate power from tracker (tracker contribution only)
}
