% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_tracker_sar}
\alias{get_tracker_sar}
\title{Calculate SAR from tracker (watch contribution only)}
\usage{
get_tracker_sar(params, tissue_params)
}
\arguments{
\item{params}{Parameter list}

\item{tissue_params}{Tissue-specific parameter list (SAR values)}
}
\value{
SAR value
}
\description{
Calculate SAR from tracker (watch contribution only)
}
