% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_tracker_tracker_dose}
\alias{get_tracker_tracker_dose}
\title{Calculate contribution from tracker}
\usage{
get_tracker_tracker_dose(duration_tracker, params, tissue_params)
}
\arguments{
\item{duration_tracker}{Daily duration of tracker use in seconds}

\item{params}{Parameter list}

\item{tissue_params}{tissue-specific parameter list}
}
\description{
Calculate contribution from tracker
}
