% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_vr_pwr}
\alias{get_vr_pwr}
\title{Calculate output power from virtual reality headset}
\usage{
get_vr_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Aggregated output power of VR headset in mW
}
\description{
Calculate output power from virtual reality headset
}
