% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_vr_sar}
\alias{get_vr_sar}
\title{Calculate SAR from VR headset}
\usage{
get_vr_sar(params, tissue_params)
}
\arguments{
\item{params}{Parameter list}

\item{tissue_params}{Tissue params}
}
\value{
tissue-specific SAR from VR headset
}
\description{
Calculate SAR from VR headset
}
