% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\name{get_wifi_pwr}
\alias{get_wifi_pwr}
\title{Calculate WiFi Aggregated Power}
\usage{
get_wifi_pwr(params)
}
\arguments{
\item{params}{parameter list}
}
\value{
aggregated output power wifi
}
\description{
Calculate WiFi Aggregated Power
}
