% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\name{get_wifi_sar}
\alias{get_wifi_sar}
\title{Calculate WiFi Aggregated SAR}
\usage{
get_wifi_sar(params, tissue_params)
}
\arguments{
\item{params}{description}

\item{tissue_params}{description}
}
\value{
description
}
\description{
Calculate WiFi Aggregated SAR
}
