% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headphones.R
\name{headphones_dose}
\alias{headphones_dose}
\title{Calculate RF-EMF Dose from Bluetooth Headphones}
\usage{
headphones_dose(tissue, duration_headphones, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_headphones}{duration of time Bluetooth headphones were used in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from Bluetooth headphone use in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from bluetooth headphones (calls excluded)
}
\details{
The RF-EMF dose from Bluetooth headphone use is calculated as:

\deqn{Dose_{headphones} = Dose_{earset} + Dose_{phone}}

Where:
\itemize{
\item \eqn{Dose_{earset}} is the dose contribution from the headphones themselves
\item \eqn{Dose_{phone}} is the dose contribution from the phone connection to the headphones
}
}
\examples{
headphones_dose(
  tissue              = "body",
  duration_headphones = 6000)

}
\seealso{
\code{\link[=headphones_earset_dose]{headphones_earset_dose()}}, \code{\link[=headphones_phone_dose]{headphones_phone_dose()}}
}
