% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{hotspot_dose}
\alias{hotspot_dose}
\title{Calculate RF-EMF Dose from Mobile Phone Hotspot}
\usage{
hotspot_dose(tissue, duration_hotspot, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_hotspot}{duration of time hotspot was active in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from hotspot use in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from hotspot
}
\details{
The RF-EMF dose from hotspot use is calculated as:

\deqn{Dose_{hotspot} = mSAR_{hotspot} * duration_{hotspot}}

Where:
\itemize{
\item \eqn{mSAR_{hotspot}} is the mSAR of hotspot use
\item \eqn{Dduration_{hotspot}} is the duration of hotspot use in s/day
}
}
\examples{
hotspot_dose(
  tissue           = "body",
  duration_hotspot = 600)

}
