% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptop.R
\name{laptop_dose}
\alias{laptop_dose}
\title{Calculate RF-EMF Dose from Laptop Use}
\usage{
laptop_dose(
  tissue,
  dur_low,
  dur_lowmed,
  dur_medhigh,
  dur_high,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{dur_low}{Duration (in seconds per day) of low output power activities on laptop}

\item{dur_lowmed}{Duration (in seconds per day) of low-medium output power activities on laptop}

\item{dur_medhigh}{Duration (in seconds per day) of medium-high output power activities on laptop}

\item{dur_high}{Duration (in seconds per day) of high output power activities on laptop}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from laptop use in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from laptop use.
}
\details{
The laptop RF-EMF dose is calculated as:

\deqn{Dose_{laptop} = mSAR_{laptop}*duration_{laptop}}

Where:
\itemize{
\item \eqn{Dose_{laptop}} is the dose from laptop use
\item \eqn{mSAR_{laptop}} is the momentary SAR value in mJ/kg
\item \eqn{duration_{laptop}} is the duration of laptop use
}

We distinguish between 4 types of activities:
\enumerate{
\item Low output power: sending e-mails, browsing the internet, scrolling and
chatting on social media, sending text messages
\item Low to medium output power: online gaming, streaming music, sending voice
messages
\item Medium to high output power: watching videos, uploading pictures or
videos, making video calls
\item High output power: uploading large files
}
}
\examples{
laptop_dose(
  tissue      = "brain",
  dur_low     = 0,
  dur_lowmed  = 1230,
  dur_medhigh = 0,
  dur_high    = 0)

}
\seealso{
\code{\link[=laptop_msar]{laptop_msar()}}
}
