% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptop.R
\name{laptop_msar}
\alias{laptop_msar}
\title{Calculate mSAR from Laptop Use}
\usage{
laptop_msar(
  tissue,
  dur_low,
  dur_lowmed,
  dur_medhigh,
  dur_high,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate mSAR (default: "brain" or "body")}

\item{dur_low}{Duration (in seconds per day) of low output power activities on laptop}

\item{dur_lowmed}{Duration (in seconds per day) of low-medium output power activities on laptop}

\item{dur_medhigh}{Duration (in seconds per day) of medium-high output power activities on laptop}

\item{dur_high}{Duration (in seconds per day) of high output power activities on laptop}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific mSAR from laptop use in mW/kg
}
\description{
Calculates the tissue-specific momentary SAR (mSAR) from laptop use.
}
\details{
The mSAR is calculated as:

\deqn{mSAR_{laptop} = nSAR_{laptop}*outputpower_{laptop}}

Where:
\itemize{
\item \eqn{mSAR_{laptop}} is the mSAR from laptop use
\item \eqn{nSAR_{laptop}} is the normalised SAR value in mJ/kg
\item \eqn{outputpower_{laptop}} is the duration of laptop use
}

We distinguish between 4 types of activities:
\enumerate{
\item Low output power: sending e-mails, browsing the internet, scrolling and
chatting on social media, sending text messages
\item Low to medium output power: online gaming, streaming music, sending voice
messages
\item Medium to high output power: watching videos, uploading pictures or
videos, making video calls
\item High output power: uploading large files
}
}
\examples{
laptop_msar(
  tissue      = "brain",
  dur_low     = 0,
  dur_lowmed  = 1230,
  dur_medhigh = 0,
  dur_high    = 0)

}
\seealso{
\code{\link[=laptop_pwr]{laptop_pwr()}}, \code{\link[=laptop_sar]{laptop_sar()}}
}
