% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptop.R
\name{laptop_pwr}
\alias{laptop_pwr}
\title{Calculate Laptop Output Power}
\usage{
laptop_pwr(
  band,
  dur_low,
  dur_lowmed,
  dur_medhigh,
  dur_high,
  params = load_params()
)
}
\arguments{
\item{band}{WiFi frequency band ("2" for 2.4 GHz, "5" for 5.0 GHz)}

\item{dur_low}{Duration (in seconds per day) of low output power activities on laptop}

\item{dur_lowmed}{Duration (in seconds per day) of low-medium output power activities on laptop}

\item{dur_medhigh}{Duration (in seconds per day) of medium-high output power activities on laptop}

\item{dur_high}{Duration (in seconds per day) of high output power activities on laptop}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power in mW
}
\description{
Calculates the laptop output power during use
}
\details{
The output power depends on the frequency band (2.4 GHz or 5.0 GHz) and the
type of activity.
}
\examples{
laptop_pwr(
  band        = "2",
  dur_low     = 0,
  dur_lowmed  = 1230,
  dur_medhigh = 0,
  dur_high    = 0)

}
