% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_default_params}
\alias{load_default_params}
\title{Load default parameter list}
\usage{
load_default_params(dest_file = NULL)
}
\arguments{
\item{dest_file}{location where downloaded YAML file should be saved}
}
\description{
Load default parameter list
}
