% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_device_params}
\alias{load_device_params}
\title{Load device-specific parameters}
\usage{
load_device_params(params, device_type)
}
\arguments{
\item{params}{Parameter list}

\item{device_type}{Device type}
}
\value{
Parameter list specific to selected device
}
\description{
Load device-specific parameters
}
