% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_tissue_params}
\alias{load_tissue_params}
\title{Load tissue-specific parameters}
\usage{
load_tissue_params(params, device_type, tissue_name)
}
\arguments{
\item{tissue_name}{Name of tissue}

\item{device_params}{Device-specific parameter list}
}
\value{
Parameter list specific to selected device and tissue
}
\description{
Load tissue-specific parameters
}
