% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{location_props}
\alias{location_props}
\title{Calculate proportion of time spent at home vs work vs outdoors vs travelling}
\usage{
location_props(travel_time, home_prop, work_prop, outd_prop)
}
\arguments{
\item{travel_time}{daily time spent travelling / commuting in seconds (s)}

\item{home_prop}{proportion of time spent at home WITHOUT considering commute}

\item{work_prop}{proportion of time spent at work WITHOUT considering commute}

\item{outd_prop}{proportion of time spent outside WITHOUT considering commute}
}
\value{
List with following elements: "travel", "home", "work", "out", corresponding
value is the proportion of time spent in each location.
}
\description{
Based on travel time (input variable by user). Assumes a fixed ratio of time
spent at home vs at work vs outdoors.
}
\details{
We assume that the time spent commuting takes away from time spent
at home, but not the time spent at work or outdoors.
}
