% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_custom_base_params}
\alias{merge_custom_base_params}
\title{Merge custom parameters with base parameters}
\usage{
merge_custom_base_params(base, custom)
}
\arguments{
\item{base}{base parameter list}

\item{custom}{custom parameter list}
}
\value{
Modified parameter YAML
}
\description{
Merge custom parameters with base parameters
}
