% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_bt_msar}
\alias{mobilecall_bt_msar}
\title{Calculate call mSAR (bluetooth contribution from phone and headphones)}
\usage{
mobilecall_bt_msar(tissue, params = load_params())
}
\arguments{
\item{tissue}{Tissue}

\item{params}{parameter list}
}
\description{
Calculate call mSAR (bluetooth contribution from phone and headphones)
}
