% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_bt_phone_pwr}
\alias{mobilecall_bt_phone_pwr}
\title{Calculate call output power (bluetooth contribution only, phone only)}
\usage{
mobilecall_bt_phone_pwr(params = load_params())
}
\arguments{
\item{params}{parameter list}
}
\description{
Calculate call output power (bluetooth contribution only, phone only)
}
