% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_bt_sar}
\alias{mobilecall_bt_sar}
\title{Calculate call sar (bluetooth contribution only, headphones only)}
\usage{
mobilecall_bt_sar(tissue, params = load_params())
}
\arguments{
\item{tissue}{brain or body}

\item{params}{parameter list}
}
\description{
Calculate call sar (bluetooth contribution only, headphones only)
}
