% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_msar}
\alias{mobilecall_msar}
\title{Calculate momentary SAR (mSAR) from mobile calling -------------------------}
\usage{
mobilecall_msar(prop_native, prop_data, prop_wifi)
}
\description{
Calculate momentary SAR (mSAR) from mobile calling -------------------------
}
