% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_bt_phone_msar}
\alias{mpc_bt_phone_msar}
\title{Calculate mobile call dose from Bluetooth headphones (phone only)}
\usage{
mpc_bt_phone_msar(tissue, headp_ear_num, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate mSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Dose from Bluetooth calls (phone contribution only) in mW/kg)
}
\description{
Calculates the RF-EMF dose from mobile calling using Bluetooth headphones
(phone contribution only; the contribution of the headphones/earset is calculated
separately).
}
\examples{
mpc_bt_phone_msar(
tissue = "brain"
)

}
\seealso{
\code{\link[=mpc_bt_earset_msar]{mpc_bt_earset_msar()}}
}
