% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_bt_phone_pwr}
\alias{mpc_bt_phone_pwr}
\title{Calculate call output power (bluetooth contribution only, phone only)}
\usage{
mpc_bt_phone_pwr(params = load_params())
}
\arguments{
\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power bluetooth headphones (phone only) in mJ
}
\description{
Description TBA
}
