% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_bt_phone_sar}
\alias{mpc_bt_phone_sar}
\title{Calculate call output power (bluetooth contribution only, headphones only)}
\usage{
mpc_bt_phone_sar(tissue, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
nSAR in W/kg/W
}
\description{
Returns the tissue-speficic nSAR from the mobile phones during calls when
connected to Bluetooth headphones (Bluetooth contribution only!)
}
