% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_bt_pwr}
\alias{mpc_bt_pwr}
\title{Calculate call output power (Bluetooth contribution only, headphones only)}
\usage{
mpc_bt_pwr(params = load_params())
}
\arguments{
\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power bluetooth headphones (headphones only) in mJ
}
\description{
Returns the output power of Bluetooth headphones during mobile phone calls.
}
