% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_bt_sar}
\alias{mpc_bt_sar}
\title{Calculate call sar (bluetooth contribution only, headphones only)}
\usage{
mpc_bt_sar(tissue, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
nSAR in W/kg/W
}
\description{
Returns the tissue-specific nSAR value from Bluetooth headphones during mobile
phone calls.
}
