% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_msar_data}
\alias{mpc_msar_data}
\title{Calculate data call mSAR}
\usage{
mpc_msar_data(
  tissue,
  headp_prop,
  ear_prop,
  speaker_prop,
  urbanicity,
  use_5g,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue}

\item{headp_prop}{Headp prop}

\item{ear_prop}{Ear prop}

\item{speaker_prop}{Speaker prop}

\item{urbanicity}{Urbanicity}

\item{use_5g}{Use 5G}

\item{travel_time}{Travel time}

\item{params}{parameter list}
}
\description{
Calculate data call mSAR
}
