% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_msar_data_bytech}
\alias{mpc_msar_data_bytech}
\title{Calculate data call mSAR by technology (2G, 3G, 4G, 5G)}
\usage{
mpc_msar_data_bytech(
  tissue,
  tech,
  headp_prop,
  ear_prop,
  speaker_prop,
  urbanicity,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue}

\item{tech}{technology (2G, 3G, 4G, 5G)}

\item{headp_prop}{Headp prop}

\item{ear_prop}{Ear prop}

\item{speaker_prop}{Speaker prop}

\item{urbanicity}{Urbanicity}

\item{travel_time}{Travel time}

\item{params}{parameter list}
}
\description{
Calculate data call mSAR by technology (2G, 3G, 4G, 5G)
}
