% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_msar_native}
\alias{mpc_msar_native}
\title{Calculate native call mSAR}
\usage{
mpc_msar_native(
  tissue,
  headp_prop,
  ear_prop,
  speaker_prop,
  urbanicity,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue}

\item{headp_prop}{Headp prop}

\item{ear_prop}{Ear prop}

\item{speaker_prop}{Speaker prop}

\item{urbanicity}{Urbanicity}

\item{travel_time}{Travel time}

\item{params}{parameter list}
}
\description{
Calculate native call mSAR
}
