% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_msar_wifi}
\alias{mpc_msar_wifi}
\title{Calculate WiFi call mSAR}
\usage{
mpc_msar_wifi(
  tissue,
  headp_prop,
  ear_prop,
  speaker_prop,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue}

\item{headp_prop}{Headp prop}

\item{ear_prop}{Ear prop}

\item{speaker_prop}{Speaker prop}

\item{params}{parameter list}
}
\description{
Calculate WiFi call mSAR
}
