% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_msar_wifi_bytech}
\alias{mpc_msar_wifi_bytech}
\title{Calculate WiFi call mSAR by band (2.4 GHz, 5.0 GHz)}
\usage{
mpc_msar_wifi_bytech(
  tissue,
  tech,
  headp_prop,
  ear_prop,
  speaker_prop,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue}

\item{tech}{technology (2.4 GHz, 5.0 GHz)}

\item{headp_prop}{Headp prop}

\item{ear_prop}{Ear prop}

\item{speaker_prop}{Speaker prop}

\item{params}{parameter list}
}
\description{
Calculate WiFi call mSAR by band (2.4 GHz, 5.0 GHz)
}
