% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_pwr_native}
\alias{mpc_pwr_native}
\title{Calculate Mobile Phone Output Power during Native Mobile Phone Calls}
\usage{
mpc_pwr_native(band, urbanicity, travel_time, params = load_params())
}
\arguments{
\item{band}{Technology (2G, 3g, 4g, or 5g) used for the call}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power in mW
}
\description{
Calculates the mobile phone output power during calls using native network.
}
\details{
The output power depends in the technology used and the location in which the
call is performed (urbanicity, indoors/outdoors/commuting)
}
\examples{
mpc_pwr_native(
band        = "4g",
urbanicity  = "suburban",
travel_time = 1800)

}
