% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{mpd_msar}
\alias{mpd_msar}
\title{Calculation of mSAR from Mobile Phone Data}
\usage{
mpd_msar(
  tissue,
  duration_low,
  duration_lowmed,
  duration_medhigh,
  duration_high,
  use_5g,
  wifi_prop_home,
  wifi_prop_work,
  wifi_prop_travel,
  urbanicity,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate mSAR (default: "brain" or "body")}

\item{duration_low}{Duration (in seconds per day) of low output power activities}

\item{duration_lowmed}{Duration (in seconds per day) of low-medium output power activities}

\item{duration_medhigh}{Duration (in seconds per day) of medium-high output power activities}

\item{duration_high}{Duration (in seconds per day) of high output power activities}

\item{use_5g}{TRUE if 5G services are used for data transfer, FALSE if not}

\item{wifi_prop_home}{Proportion of time connected to WiFi (vs mobile data)
at home}

\item{wifi_prop_work}{Proportion of time connected to WiFi (vs mobile data)
at school / work}

\item{wifi_prop_travel}{Proportion of time connected to WiFi (vs mobile data)
while commuting}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Momentary SAR (mSAR) from mobile data use in mW/kg
}
\description{
Calculates mSAR (momentary specifc absorption rate) from data (mobile data
and WiFi) use for a specific tissue
}
\details{
The mSAR is calculated as

\deqn{mSAR = nSAR_{mobiledata} \times
outputpower_{mobiledata} + nSAR_{wifi} \times outputpower_{wifi}}

where
\itemize{
\item \eqn{nSAR_{data}}, \eqn{nSAR_{wifi}} are the nSAR
(normalized specific absorption rates, in W/kg/W) from data use using
mobile data or WiFi networks, respectively
\item \eqn{outputpower_{data}}, \eqn{outputpower_{wifi}} are the output powers (mW)
of the device (mobilephone) using mobile data or WiFi networks, respectively
}
}
\examples{
mpd_msar(
tissue           = "brain",
duration_low     = 4860,
duration_lowmed  = 540,
duration_medhigh = 4860,
duration_high    = 540,
use_5g           = FALSE,
wifi_prop_home   = 0.5,
wifi_prop_work   = 0.5,
wifi_prop_travel = 0.5,
urbanicity       = "suburban",
travel_time      = 1800)

}
\seealso{
\code{\link[=mpd_pwr_data(), mpd_pwr_wifi(), mpd_sar_data(), mpd_sar_wifi]{mpd_pwr_data(), mpd_pwr_wifi(), mpd_sar_data(), mpd_sar_wifi()}}
}
