% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{mpd_pwr_wifi}
\alias{mpd_pwr_wifi}
\title{Calculate Mobile Phone Output Power during WiFi Use}
\usage{
mpd_pwr_wifi(
  band,
  duration_low,
  duration_lowmed,
  duration_medhigh,
  duration_high,
  params = load_params()
)
}
\arguments{
\item{band}{WiFi frequency band ("2" for 2.4 GHz, "5" for 5.0 GHz)}

\item{duration_low}{Duration (in seconds per day) of low output power activities}

\item{duration_lowmed}{Duration (in seconds per day) of low-medium output power activities}

\item{duration_medhigh}{Duration (in seconds per day) of medium-high output power activities}

\item{duration_high}{Duration (in seconds per day) of high output power activities}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power in mW
}
\description{
Calculates the mobile phone output power during WiFi use
}
\details{
The output power depends on the frequency band (2.4 GHz or 5.0 GHz) and the
type of activity.
}
\examples{
mpd_pwr_wifi(
band             = "2",
duration_low     = 4860,
duration_lowmed  = 540,
duration_medhigh = 4860,
duration_high    = 540)

}
