% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{mpd_sar_data}
\alias{mpd_sar_data}
\title{Calculate nSAR during mobile data use on mobile phone}
\usage{
mpd_sar_data(band, tissue, params = load_params())
}
\arguments{
\item{band}{Technology ("3g", "4g", or "5g")}

\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
nSAR in W/kg/W
}
\description{
Calculates tissue-specific nSAR from mobile data use on mobile phone.
}
\details{
The normalized specific absorption rate (nSAR) depends on the tissue and
the frequency band.
}
\examples{
mpd_sar_data(
tissue       = "brain",
band         = "5g",
params       = load_params())

}
