% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nested_params_to_flat}
\alias{nested_params_to_flat}
\title{Convert nested parameter YAML to flat parameter table}
\usage{
nested_params_to_flat(nested, parent_path = NULL)
}
\arguments{
\item{nested}{nested YAML parameter list}

\item{parent_path}{parent path (optional)}
}
\value{
flat parameter list
}
\description{
Convert nested parameter YAML to flat parameter table
}
