% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{recode_urbanicity}
\alias{recode_urbanicity}
\title{Check and re-code input values - urbanicity}
\usage{
recode_urbanicity(urbanicity)
}
\arguments{
\item{urbanicity}{descr}
}
\value{
re-coded urbanicity (binary variables)
}
\description{
Check and re-code input values - urbanicity
}
