% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_default_params_file}
\alias{save_default_params_file}
\title{Load default parameter list}
\usage{
save_default_params_file(dest_file = NULL)
}
\arguments{
\item{dest_file}{location where downloaded YAML file should be saved}
}
\description{
Load default parameter list
}
