% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{scale_power_by_use5g}
\alias{scale_power_by_use5g}
\title{Scale band power by 5G use TRUE/FALSE ---------------------------------------}
\usage{
scale_power_by_use5g(use_5g, pwr_3g, pwr_4g, pwr_5g, params)
}
\arguments{
\item{use_5g}{...}

\item{pwr_3g}{...}

\item{pwr_4g}{...}

\item{pwr_5g}{...}

\item{params}{...}
}
\description{
Scale band power by 5G use TRUE/FALSE ---------------------------------------
}
