% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartwatch.R
\name{smartwatch_dose}
\alias{smartwatch_dose}
\title{Calculate RF-EMF Dose from Smart Watch}
\usage{
smartwatch_dose(tissue, duration_smartwatch, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_smartwatch}{duration of time smartwatch was worn in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from smart-watch use in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from smart-watch use
}
\details{
The RF-EMF dose from smart-watch use is calculated as:

\deqn{Dose_{smartwatch} = Dose_{watch} + Dose_{phone}}

Where:
\itemize{
\item \eqn{Dose_{watch}} is the dose contribution from the smart-watch itself
\item \eqn{Dose_{phone}} is the dose contribution from the phone connection to the smart-watch
}
}
\examples{
smartwatch_dose(
  tissue              = "body",
  duration_smartwatch = 6000)

}
\seealso{
\code{\link[=smartwatch_watch_dose]{smartwatch_watch_dose()}}, \code{\link[=smartwatch_phone_dose]{smartwatch_phone_dose()}}
}
