% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartwatch.R
\name{smartwatch_phone_dose}
\alias{smartwatch_phone_dose}
\title{Calculate RF-EMF Dose from Smart Watch (phone contribution only)}
\usage{
smartwatch_phone_dose(tissue, duration_smartwatch, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_smartwatch}{duration of time smart-watch was worn in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from smart-watch use (phone contribution only) in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from smart-watch use (phone
contribution only).
}
\details{
The RF-EMF dose from smart-watch use (phone contribution) is calculated as:

\deqn{Dose_{phone} = mSAR_{phone} * duration_{phone}}

Where:
\itemize{
\item \eqn{mSAR_{phone}} is the momentary SAR
\item \eqn{duration_{phone}} is the use duration
}
}
\examples{
smartwatch_phone_dose(
  tissue              = "body",
  duration_smartwatch = 6000)

  @export
}
