% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total_dose}
\alias{total_dose}
\title{Calculate Total RF-EMF Dose from All Sources}
\usage{
total_dose(sample, tissue, params = load_params())
}
\arguments{
\item{sample}{A list with input values for a single observation. List elements must
match the names and data types of the provided example dataset and must not be NA.}

\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Total RF-EMF dose in mJ/kg/day for the specified tissue.
}
\description{
Calculates the total RF-EMF dose from all available RF-EMF sources
for a specific tissue.
}
\details{
A detailed description of the dose calculation will follow.
}
