% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_reality.R
\name{vr_dose}
\alias{vr_dose}
\title{Calculate RF-EMF Dose from Virtual Reality Headset}
\usage{
vr_dose(tissue, duration_vr, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_vr}{duration of time VR headset was used in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from VR headset use in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from VR headset
}
\details{
The RF-EMF dose from VR headset is calculated as:

\deqn{Dose_{vr} = mSAR_{vr} * duration_{vr}}

Where:
\itemize{
\item \eqn{mSAR_{vr}} is the mSAR of virtual reality headet use
\item \eqn{Dduration_{vr}} is the duration of VR headset use in s/day
}
}
\examples{
vr_dose(
  tissue      = "brain",
  duration_vr = 600)

}
