% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\name{wifi_dose}
\alias{wifi_dose}
\title{Calculate RF-EMF Dose from WiFi Router}
\usage{
wifi_dose(tissue, travel_time, wifi_prop_travel, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{travel_time}{Time spent commuting in seconds per day}

\item{wifi_prop_travel}{Proportion of time connected to WiFi (vs mobile data)
while commuting}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from Wifi router in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from WiFi router
}
\details{
The RF-EMF dose from WiFi router is calculated as:

\deqn{Dose_{WiFi} = mSAR_{farfield}*duration_{WiFi}}

Where:
\itemize{
\item \eqn{Dose_{WiFi}} is the dose from WiFi Router
\item \eqn{mSAR_{WiFi}} is the momentary SAR value in mJ/kg
\item \eqn{duration_{WiFi}} is the duration of time exposed to a WiFi router in seconds/day
}
}
\examples{
wifi_dose(
  tissue           = "body",
  travel_time      = 1800,
  wifi_prop_travel = 0.5)

}
\seealso{
\code{\link[=wifi_msar]{wifi_msar()}}
}
