% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\name{wifi_pwr}
\alias{wifi_pwr}
\title{Calculate Output Power of WiFi Router}
\usage{
wifi_pwr(band, params = load_params())
}
\arguments{
\item{band}{Frequency band ("2" for 2.4 GHz, "5" for 5.0 GHz)}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power in mW/m**2
}
\description{
Calculates the output power of WiFi router
}
\details{
The output power depends on the WiFi frequency band.
}
\examples{
wifi_pwr(
band = "2")

}
