% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{act_pwr_props}
\alias{act_pwr_props}
\title{Get low/lowmed/medhigh/high activity power proportions}
\usage{
act_pwr_props(low_dur, lowmed_dur, medhigh_dur, high_dur)
}
\arguments{
\item{low_dur}{...}

\item{lowmed_dur}{...}

\item{medhigh_dur}{...}

\item{high_dur}{...}
}
\description{
Get low/lowmed/medhigh/high activity power proportions
}
