% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{act_pwr_props}
\alias{act_pwr_props}
\title{Get low/lowmed/medhigh/high activity power proportions}
\usage{
act_pwr_props(low_dur, lowmed_dur, medhigh_dur, high_dur)
}
\arguments{
\item{low_dur}{Duration (in seconds per day) of low output power activities}

\item{lowmed_dur}{Duration (in seconds per day) of low-medium output power activities}

\item{medhigh_dur}{Duration (in seconds per day) of medium-high output power activities}

\item{high_dur}{Duration (in seconds per day) of high output power activities}
}
\value{
Named list with the use proportion of each activity.
}
\description{
Converts activity durations to proportions.
}
