% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_path}
\alias{add_path}
\title{Helper function to convert flat parameter list to nested YAML}
\usage{
add_path(lst, keys, value)
}
\value{
nested parameter list
}
\description{
Helper function to convert flat parameter list to nested YAML
}
