% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_data_tech_proportions}
\alias{calculate_data_tech_proportions}
\title{Get mobile data technology use proportions}
\usage{
calculate_data_tech_proportions(use_5g)
}
\arguments{
\item{use_5g}{boolean variable (TRUE if participant uses 5G, FALSE otherwise)}
}
\value{
list with technology use proportions
}
\description{
Based on user variable use_5g that indicates if participant uses 5G on mobile
phone or not.
}
