% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{calculate_emf_doses}
\alias{calculate_emf_doses}
\title{Calculate Total RF-EMF Doses for ALL samples}
\usage{
calculate_emf_doses(data, param_file = NULL, default_value_file = NULL)
}
\arguments{
\item{data}{A data frame. Must have the following columns: DOC TO BE ADDED SOON}

\item{param_file}{optional path to YAML parameter configuration file}

\item{default_value_file}{optional path to YAML default value file in case of missing data}
}
\value{
A data frame. Columns named SOURCE_dose_TISSUE contain RF-EMF dose
of each participant in mJ/kg/day
}
\description{
Wrapper function that calculates doses for each row of a data frame.
}
