% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{calculate_emf_doses}
\alias{calculate_emf_doses}
\title{Calculate Total RF-EMF Doses for ALL samples}
\usage{
calculate_emf_doses(data)
}
\arguments{
\item{data}{A data frame. Must have the following columns:
sample_id: unique participant identifier
mpc_duration: daily duration of mobile phone call in seconds
mpc_ear_prop: proportion of time phone is held against ear during call
urbanicity: urban, suburban, or rural
mpd_duration: daily duration of mobile data traffic in seconds
mpd_wifi_prop: proportion of time WiFi is used (vs data) for mobile data traffic
dect_duration: daily duration of DECT/cordless phone call in seconds
dect_ear_prop:
lptp_duration:
tblt_duration:
wifi_duration:}
}
\value{
A data frame. Columns named SOURCE_dose_TISSUE contain RF-EMF dose
of each participant in mJ/kg/day
}
\description{
Calculate Total RF-EMF Doses for ALL samples
}
