% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{calculate_emf_doses}
\alias{calculate_emf_doses}
\title{Calculate Total RF-EMF Doses for ALL samples}
\usage{
calculate_emf_doses(data, tissue, params = NULL, default_value_file = NULL)
}
\arguments{
\item{data}{A data frame. Must have the exact same columns as the provided example dataset.
Columns must not be missing. Missing values will be replaced with default values.}

\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{params}{(experimental) Path to a parameter file in .yaml format. Must match the structure
of the internal parameter file. Optional; if not specified, calculations use
default parameters.}

\item{default_value_file}{List of default values to replace missing values (optional).}
}
\value{
A data frame. Columns contain the calculated RF-EMF dose
of each row in mJ/kg/day, for each exposure source and each tissue.
}
\description{
Wrapper function that calculates RF-EMF doses for a data frame of observations.
This function will be removed in the future.
}
\details{
Additional details will follow.
}
