% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{calculate_emf_doses}
\alias{calculate_emf_doses}
\title{Calculate Total RF-EMF Doses for ALL samples}
\usage{
calculate_emf_doses(
  data,
  tissue,
  params = load_params(),
  default_value_file = NULL
)
}
\arguments{
\item{data}{A data frame with required columns.}

\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{default_value_file}{Optional path to external YAML default value file. Must follow same structure as internal YAML default value file.}

\item{param_file}{Optional path to an external YAML parameter file. Must follow same structure as internal YAML parameter file.}
}
\value{
A data frame. Columns named SOURCE_dose_TISSUE contain the calculated RF-EMF dose
of each row in mJ/kg/day, for each exposure source and each tissue.
}
\description{
Wrapper function that calculates RF-EMF doses for a full data frame.
}
