% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{calculate_emf_doses}
\alias{calculate_emf_doses}
\title{Calculate Total RF-EMF Doses for ALL samples}
\usage{
calculate_emf_doses(data)
}
\arguments{
\item{test}{a data frame}
}
\value{
a data frame
}
\description{
Calculate Total RF-EMF Doses for ALL samples
}
