% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_boolean_not_na}
\alias{check_boolean_not_na}
\title{Check if value is boolean and non-NA}
\usage{
check_boolean_not_na(x)
}
\arguments{
\item{x}{number}

\item{name}{name of value}
}
\description{
Check if value is boolean and non-NA
}
