% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_character_not_na}
\alias{check_character_not_na}
\title{Check if value is character and non-NA}
\usage{
check_character_not_na(x, name)
}
\arguments{
\item{x}{number}

\item{name}{name of value}
}
\description{
Check if value is character and non-NA
}
