% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_character_not_na}
\alias{check_character_not_na}
\title{Check if value is character and non-NA}
\usage{
check_character_not_na(x)
}
\arguments{
\item{x}{Single value to be checked: should be character, atomic, and not NA
to pass check}
}
\value{
Returns invisible TRUE if check passes. Side effect of stopping
calculation if check fails.
}
\description{
Check if value is character and non-NA
}
