% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_duration}
\alias{check_duration}
\title{Check input values - duration}
\usage{
check_duration(duration)
}
\arguments{
\item{duration}{duration}
}
\value{
TRUE if duration is valid, FALSE if duration is not valid
}
\description{
Check input values - duration
}
