% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_headp_num}
\alias{check_headp_num}
\title{Check input values - number of headphones}
\usage{
check_headp_num(headp_num)
}
\arguments{
\item{headp_num}{Number of headphones}
}
\description{
Check input values - number of headphones
}
