% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_proportions}
\alias{check_proportions}
\title{Check input values - proportions}
\usage{
check_proportions(proportions)
}
\arguments{
\item{proportions}{Vector with proportions to check}
}
\value{
TRUE if proportions are valid, FALSE if invalid
}
\description{
Check input values - proportions
}
