% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_tissue}
\alias{check_tissue}
\title{Check tissue}
\usage{
check_tissue(tissue, device, params)
}
\arguments{
\item{tissue}{Tissue}

\item{device}{Device}

\item{params}{Params}
}
\description{
Check tissue
}
