% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordless.R
\name{dect_msar}
\alias{dect_msar}
\title{Calculate mSAR from Cordless Calls}
\usage{
dect_msar(tissue, ear_prop, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{ear_prop}{Proportion of time cordless phone is held against ear
during calls.}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific mSAR from DECT/cordless phone calls in mW/kg
}
\description{
Calculates mSAR (momentary SAR) from making calls with cordless phone (also known as
DECT phone).
}
\details{
The mSAR is calculated as

\deqn{mSAR_{DECT} = nSAR_{DECT}*outputpower_{DECT}}

Where:
\itemize{
\item \eqn{mSAR_{DECT}} is the mSAR from DECT/cordless phone calls
\item \eqn{nSAR_{DECT}} is the nSAR (normalized specific absorption rate) in W/kg/W
\item \eqn{outputpower_{DECT}} is the duration of the calls in mW
}
}
\examples{
dect_msar(
  tissue      = "brain",
  ear_prop    = 0.9)

}
\seealso{
\code{\link[=dect_pwr(), dect_sar]{dect_pwr(), dect_sar()}}
}
