% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data for testing}
\format{
A data frame with 50 rows
\describe{
  \item{dect_duration}{test}
  \item{dect_ear_prop}{test}
  \item{gaming_duration}{test}
  \item{headp_ear_num}{test}
  \item{headphone_duration}{test}
  \item{hotspot_duration}{test}
  \item{lptp_duration}{test}
  \item{mpc_duration}{test}
  \item{mpc_ear_prop}{test}
  \item{mpc_headp_prop}{test}
  \item{mpd_duration}{test}
  \item{mpd_high_dt_dur}{test}
  \item{mpd_high_dt_prop}{test}
  \item{mpd_low_dt_dur}{test}
  \item{mpd_lowmed_dt_dur}{test}
  \item{mpd_medhigh_dt_dur}{test}
  \item{mpd_wifi_prop}{test}
  \item{sample_id}{test}
  \item{smarthome_duration}{test}
  \item{smartwatch_duration}{test}
  \item{tblt_duration}{test}
  \item{tracker_duration}{test}
  \item{travel_time}{test}
  \item{urbanicity}{test}
  \item{use_5g}{test}
  \item{vr_duration}{test}
  \item{wifi_duration}{test}
}
}
\source{
Generated for package testing.
}
\usage{
example_data
}
\description{
A dataset containing example user data (made up)
}
\keyword{datasets}
