% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data for testing}
\format{
A data frame with 50 rows
\describe{
  \item{var1}{test}
  \item{var2}{test}
}
}
\source{
Generated for package testing.
}
\usage{
example_data
}
\description{
A dataset containing example user data (made up)
}
\keyword{datasets}
