% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{farfield_sar}
\alias{farfield_sar}
\title{Calculate nSAR from Far-Field Sources}
\usage{
farfield_sar(tissue, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
nSAR in W/kg/W/m**2
}
\description{
Calculates tissue-specific nSAR from far-field sources.
}
\details{
The normalized specific absorption rate (nSAR) depends on the tissue.
}
\examples{
farfield_sar(
tissue       = "brain",
params       = load_params())

}
