% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill_missing_variables}
\alias{fill_missing_variables}
\title{Fill missing variables}
\usage{
fill_missing_variables(data, defaults, warn_threshold = 0.2)
}
\arguments{
\item{data}{data}

\item{defaults}{defaults}

\item{warn_threshold}{threshold proportion of missing data to raise warning}
}
\value{
data frame with missing data replaced with default values
}
\description{
Fill missing variables
}
