% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordless.R
\name{get_cordless_dose}
\alias{get_cordless_dose}
\title{Calculate Dose from Cordless Calling}
\usage{
get_cordless_dose(duration, ear_proportion, params = NULL)
}
\arguments{
\item{duration}{Duration of cordless calls in seconds. >= 0 and <= 86400}

\item{ear_proportion}{Proportion of time cordless phone is held against ear
during calls. >=0 and <=1.}

\item{params}{Parameter list.}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from Cordless Calling
}
