% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordless.R
\name{get_cordless_sar}
\alias{get_cordless_sar}
\title{Calculate Aggregated SAR during cordless call (tissue-specific)}
\usage{
get_cordless_sar(ear_proportion, params, tissue_params)
}
\arguments{
\item{ear_proportion}{Proportion of holding cordless phone against ear}

\item{params}{descr}

\item{tissue_params}{descr}
}
\value{
sar
}
\description{
Calculate Aggregated SAR during cordless call (tissue-specific)
}
