% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{get_farfield_dose}
\alias{get_farfield_dose}
\title{Calculate Dose from far-field exposure}
\usage{
get_farfield_dose(urbanicity, travel_time, params)
}
\arguments{
\item{urbanicity}{Urbanicity of home environment}

\item{travel_time}{Time per day spent commuting (s)}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from far-field exposure
}
