% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{get_farfield_pwr}
\alias{get_farfield_pwr}
\title{Calculate far-field exposure aggregated power}
\usage{
get_farfield_pwr(urbanicity, travel_time, params)
}
\arguments{
\item{urbanicity}{Urbanicity of home environment}

\item{travel_time}{Time per day spent commuting (s)}

\item{params}{Parameter list}
}
\value{
far-field power in mw/m**2
}
\description{
Calculates far-field source output power
}
