% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{get_farfield_pwr}
\alias{get_farfield_pwr}
\title{Calculate far-field exposure aggregated power}
\usage{
get_farfield_pwr(urb_list, travel_time, params)
}
\arguments{
\item{urb_list}{list of binary urbanicitiy values created with recode_urbanicity
function}

\item{params}{description}
}
\value{
far-field power
}
\description{
Calculate far-field exposure aggregated power
}
