% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farfield.R
\name{get_farfield_sar}
\alias{get_farfield_sar}
\title{Calculate far-field exposure SAR}
\usage{
get_farfield_sar(params, tissue_params)
}
\arguments{
\item{params}{Parameter list}

\item{tissue_params}{Tissue-specific parameter list (SAR values)}
}
\value{
aggregated sar in W/kg/W/m**2
}
\description{
This function will be expanded in the future to provide a more fine-tuned
calculation of the far-field SAR values.
}
