% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_headphone_dose}
\alias{get_headphone_dose}
\title{Calculate brain and body dose from using bluetooth headphones}
\usage{
get_headphone_dose(duration_headphones, params = NULL)
}
\arguments{
\item{duration_headphones}{Daily duration of using bt headphones connected to phone}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from bt headphones connected to phone
}
\description{
Important: exposure comes from both mobile phone and from smart watch.
}
