% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptop.R
\name{get_laptop_dose}
\alias{get_laptop_dose}
\title{Calculate Dose from Laptop Use}
\usage{
get_laptop_dose(duration, params)
}
\arguments{
\item{duration}{Duration of laptop use in seconds}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from Laptop Use
}
