% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptop.R
\name{get_laptop_pwr}
\alias{get_laptop_pwr}
\title{Calculate Laptop Aggregated Power}
\usage{
get_laptop_pwr(params)
}
\arguments{
\item{params}{text}
}
\value{
text
}
\description{
Calculate Laptop Aggregated Power
}
