% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_bt_headp_pwr}
\alias{get_mobilecall_bt_headp_pwr}
\title{Calculate output power from mobile calling using Bluetooth headphones (headphone contribution)}
\usage{
get_mobilecall_bt_headp_pwr(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Output power from mobile phone (Bluetooth contribution) during mobile phone call in mJ
}
\description{
Calculates the output power from mobile calling using Bluetooth headphones (headphone Bluetooth contribution).
}
\details{
In this version, this function serves as a placeholder and simply fetches the corresponding parameter from the parameter list.
}
