% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_bt_headp_sar}
\alias{get_mobilecall_bt_headp_sar}
\title{Calculate SAR from mobile calling using Bluetooth headphones (headphone contribution)}
\usage{
get_mobilecall_bt_headp_sar(params, tissue_params)
}
\arguments{
\item{params}{Parameter list}

\item{tissue_params}{Tissue parameter list}
}
\value{
Output power from mobile phone (Bluetooth contribution) during mobile phone call in mW
}
\description{
Calculates SAR from mobile calling using Bluetooth headphones (headphone Bluetooth contribution).
}
\details{
In this version, this function serves as a placeholder and simply fetches the corresponding parameter from the parameter list.
}
