% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_bt_phone_sar}
\alias{get_mobilecall_bt_phone_sar}
\title{Mobilecall bluetooth phone sar}
\usage{
get_mobilecall_bt_phone_sar(params, tissue_params)
}
\arguments{
\item{params}{params}

\item{tissue_params}{tissue params}
}
\description{
Mobilecall bluetooth phone sar
}
