% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_dose}
\alias{get_mobilecall_dose}
\title{Calculate Dose from mobile call
This is a test equation:
\deqn{a + b}}
\usage{
get_mobilecall_dose(duration, ear_prop, urbanicity, params)
}
\arguments{
\item{duration}{Duration of mobile call in seconds}

\item{ear_prop}{Proportion of time mobile phone is held against the ear}

\item{urbanicity}{Urbanicity of environment (rural, suburban, urban)}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from mobile call
This is a test equation:
\deqn{a + b}
}
