% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_dose}
\alias{get_mobilecall_phone_dose}
\title{Mobilecall Phone dose}
\usage{
get_mobilecall_phone_dose(
  duration,
  ear_prop,
  speaker_prop,
  headp_prop,
  urbanicity,
  params,
  tissue_params
)
}
\arguments{
\item{duration}{User duration}

\item{ear_prop}{Ear proportion}

\item{speaker_prop}{Speaker proportion}

\item{headp_prop}{Headphone proportion}

\item{urbanicity}{Urbanicity}

\item{params}{Parameter list}

\item{tissue_params}{Tissue-specific parameter list}
}
\value{
Dose
}
\description{
Mobilecall Phone dose
}
