% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_pwr}
\alias{get_mobilecall_phone_pwr}
\title{Mobilecall Phone Power}
\usage{
get_mobilecall_phone_pwr(urbanicity, params)
}
\arguments{
\item{urbanicity}{Urbanicity}

\item{params}{Parameter list}
}
\value{
Power
}
\description{
Mobilecall Phone Power
}
