% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_sar}
\alias{get_mobilecall_phone_sar}
\title{Mobilecall SAR}
\usage{
get_mobilecall_phone_sar(
  ear_prop,
  headp_prop,
  speaker_prop,
  params,
  tissue_params
)
}
\arguments{
\item{ear_prop}{Ear proportion}

\item{headp_prop}{Headphone proportion}

\item{speaker_prop}{Speaker proportion}

\item{params}{Parameter list}

\item{tissue_params}{SAR values}
}
\description{
Mobilecall SAR
}
