% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_wifi_pwr}
\alias{get_mobilecall_phone_wifi_pwr}
\title{Wifi Phone Power}
\usage{
get_mobilecall_phone_wifi_pwr(
  wifi_2_prop,
  wifi_5_prop,
  wifi_2_pwr,
  wifi_5_pwr,
  wifi_2_duty_cycle,
  wifi_5_duty_cycle
)
}
\arguments{
\item{wifi_2_prop}{Proportion of 2.4GHz WiFi}

\item{wifi_5_prop}{Proportion of 5.0GHz WiFi}

\item{wifi_2_pwr}{2.4GHz WiFi Call Power}

\item{wifi_5_pwr}{5.0GHz WiFi Call Power}

\item{wifi_2_duty_cycle}{2.4GHz WiFi Duty Cycle}

\item{wifi_5_duty_cycle}{5.0GHz WiFi Duty Cycle}
}
\description{
Wifi Phone Power
}
