% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{get_mobilecall_phone_wifi_pwr}
\alias{get_mobilecall_phone_wifi_pwr}
\title{Calculate mobile phone output power from WiFi mobile calling (no Bluetooth)}
\usage{
get_mobilecall_phone_wifi_pwr(
  wifi_2_prop,
  wifi_5_prop,
  wifi_2_pwr,
  wifi_5_pwr,
  wifi_2_duty_cycle,
  wifi_5_duty_cycle
)
}
\arguments{
\item{wifi_2_prop}{Proportion of 2.4GHz WiFi}

\item{wifi_5_prop}{Proportion of 5.0GHz WiFi}

\item{wifi_2_pwr}{2.4GHz WiFi Call Power}

\item{wifi_5_pwr}{5.0GHz WiFi Call Power}

\item{wifi_2_duty_cycle}{2.4GHz WiFi Duty Cycle}

\item{wifi_5_duty_cycle}{5.0GHz WiFi Duty Cycle}
}
\value{
Mobile phone output power during WiFi mobile calling (no Bluetooth) in mW
}
\description{
Calculates the output power of the mobile phone during WiFi mobile calls in mW (No Bluetooth)
}
\details{
The mobile phone output power during WiFi mobile calling (no Bluetooth) is calculated as:
\deqn{Power_{WiFi} = Prop_{2.4GHz} \times Power_{2.4GHz} \times DutyCycle_{2.4GHz} + Prop_{5.0GHz} \times Power_{5.0GHz} \times DutyCycle_{5.0GHz}}

Where:
\itemize{
\item \eqn{Prop_{2.4GHz}}, \eqn{Prop_{5.0GHz}} are the proportion of 2.4GHz and 5.0GHz WiFi, respectively
\item \eqn{Power_{2.4GHz}}, \eqn{Power_{5.0GHz}} are the output power of 2.4GHz and 5.0GHz WiFi calls from the mobile phone (no Bluetooth), respectively
\item \eqn{DutyCycle_{2.4GHz}}, \eqn{DutyCycle_{5.0GHz}} are the duty cycle of 2.4GHz and 5.0GHz WiFi, respectively
}
}
