% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_dose}
\alias{get_mobiledata_dose}
\title{Calculate Dose from mobile data}
\usage{
get_mobiledata_dose(
  duration,
  use_5g,
  wifi_prop,
  urbanicity,
  act_pwr_props,
  travel_time,
  high_pwr_prop,
  params
)
}
\arguments{
\item{duration}{Duration of mobile data transfer in seconds}

\item{use_5g}{If participant uses 5G or not}

\item{wifi_prop}{Proportion of time WiFi connection is used for data
transfer (vs mobile data)}

\item{urbanicity}{urbanicity}

\item{act_pwr_props}{activity power proportions}

\item{travel_time}{travel time}

\item{high_pwr_prop}{Proportion of time spent on high data transfer activities}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from mobile data
}
