% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mobiledata_sar}
\alias{get_mobiledata_sar}
\title{Calculate mobile data SAR}
\usage{
get_mobiledata_sar(wifi_prop, data_prop, params, tissue_params)
}
\arguments{
\item{wifi_prop}{wifi proportion}

\item{data_prop}{data_proportion}

\item{params}{parameter list}

\item{tissue_params}{tissue-specific parameters (SAR-values)}
}
\description{
Calculate mobile data SAR
}
