% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{get_mpd_data_pwr}
\alias{get_mpd_data_pwr}
\title{Calculate mpd power from data}
\usage{
get_mpd_data_pwr(use_5g, urbanicity, act_pwr_props, loc_props, params)
}
\arguments{
\item{use_5g}{TRUE if participant uses 5G, FALSE otherwise}

\item{urbanicity}{urbanicity of home/work environment}

\item{act_pwr_props}{...}

\item{loc_props}{location proportions}

\item{params}{device-specific parameters}
}
\value{
pwr
}
\description{
Calculate mpd power from data
}
