% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_other_dose}
\alias{get_other_dose}
\title{Calculate Dose from other exposure sources}
\usage{
get_other_dose(
  duration_hotspot,
  duration_smartwatch,
  duration_tracker,
  duration_vr,
  duration_headphones,
  duration_smarthome,
  duration_gaming,
  params
)
}
\arguments{
\item{duration_hotspot}{daily hotspot use duration (s)}

\item{duration_smartwatch}{daily smart watch use duration (s)}

\item{duration_tracker}{daily tracker use duration (s)}

\item{duration_vr}{daily vr use duration (s)}

\item{duration_headphones}{daily bluetooth headphone use duration (s)}

\item{duration_smarthome}{daily time spend in smarthome (s)}

\item{duration_gaming}{daily portable gaming console use duration (s)}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from other exposure sources
}
