% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_other_dose}
\alias{get_other_dose}
\title{Calculate Dose from other exposure sources}
\usage{
get_other_dose()
}
\arguments{
\item{dur_tablet}{Duration of other exposure sources in seconds}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from other exposure sources
}
