% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smarthome_dose}
\alias{get_smarthome_dose}
\title{Calculate brain and body dose from smart home}
\usage{
get_smarthome_dose(smarthome, travel_time, params = NULL)
}
\arguments{
\item{smarthome}{TRUE if participant lives in smarthome, FALSE if not}

\item{travel_time}{Time spent commuting in s/day}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from smart home
}
\description{
Calculate brain and body dose from smart home
}
