% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_smartwatch_watch_dose}
\alias{get_smartwatch_watch_dose}
\title{Calculate contribution from smart watch}
\usage{
get_smartwatch_watch_dose(duration_smartwatch, params, tissue_params)
}
\arguments{
\item{duration_smartwatch}{Daily duration of smartwatch use in seconds}

\item{params}{Tissue parameter list}
}
\description{
Calculate contribution from smart watch
}
