% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{get_tablet_dose}
\alias{get_tablet_dose}
\title{Calculate Dose from Tablet Use}
\usage{
get_tablet_dose(duration, params)
}
\arguments{
\item{duration}{Duration of tablet use in seconds}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from Tablet Use
}
