% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{get_tablet_pwr}
\alias{get_tablet_pwr}
\title{Calculate Tablet Aggregated Power}
\usage{
get_tablet_pwr(act_pwr_props, params)
}
\arguments{
\item{act_pwr_props}{descr}

\item{params}{descr}
}
\value{
Aggregated tablet power
}
\description{
Calculate Tablet Aggregated Power
}
