% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{get_total_dose}
\alias{get_total_dose}
\title{Calculate Total RF-EMF Dose for Brain and Body from All Sources}
\usage{
get_total_dose(sample, params = NULL)
}
\arguments{
\item{sample}{A list with input values for a single sample}

\item{params}{A parameter list}
}
\value{
A list with results for brain and body dose for a single sample
}
\description{
Calculate Total RF-EMF Dose for Brain and Body from All Sources
}
