% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_tracker_dose}
\alias{get_tracker_dose}
\title{Calculate brain and body dose from using personal tracker connected to mobile phone}
\usage{
get_tracker_dose(duration_tracker, params)
}
\arguments{
\item{duration_tracker}{Daily duration of using tracker in seconds}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from tracker use
}
\description{
Important: exposure comes from both mobile phone and from smart watch.
}
