% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othersources.R
\name{get_vr_dose}
\alias{get_vr_dose}
\title{Calculate brain and body dose from using VR headset}
\usage{
get_vr_dose(duration_vr, params)
}
\arguments{
\item{duration_vr}{Daily duration of using VR headset in seconds}

\item{params}{Parameter list}
}
\value{
list with daily brain and body dose from VR headset
}
\description{
Calculate brain and body dose from using VR headset
}
