% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\name{get_wifi_dose}
\alias{get_wifi_dose}
\title{Calculate Dose from wifi Use}
\usage{
get_wifi_dose(duration, params)
}
\arguments{
\item{duration}{Duration of exposure to WiFi router}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from wifi Use
}
