% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\name{get_wifi_dose}
\alias{get_wifi_dose}
\title{Calculate Dose from wifi Use}
\usage{
get_wifi_dose(travel_time, wifi_prop_travel, params = NULL)
}
\arguments{
\item{travel_time}{Daily time spent commuting (s)}

\item{wifi_prop_travel}{Proportion of time connected to WiFi (vs mobile data) during commute}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from wifi Use
}
