% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headphones.R
\name{headphones_earset_dose}
\alias{headphones_earset_dose}
\title{Calculate RF-EMF Dose from Smart Watch (earset contribution only)}
\usage{
headphones_earset_dose(tissue, duration_headphones, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}

\item{duration_smartwatch}{duration of time Bluetooth headphones were worn in s/day}
}
\value{
Tissue-specific RF-EMF dose from Bluetooth headphone use (earset contribution only) in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from Bluetooth headphone use (earset
contribution only).
}
\details{
The RF-EMF dose from Bluetooth headphone use (earset
contribution) is calculated as:

\deqn{Dose_{earset} = mSAR_{earset} * duration_{earset}}

Where:
\itemize{
\item \eqn{mSAR_{earset}} is the momentary SAR for the headphone connection
\item \eqn{duration_{earset}} is the use duration in s/day
}
}
\examples{
headphones_earset_dose(
  tissue              = "body",
  duration_headphones = 6000)

  @export
}
