% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headphones.R
\name{headphones_phone_dose}
\alias{headphones_phone_dose}
\title{Calculate RF-EMF Dose from Smart Watch (phone contribution only)}
\usage{
headphones_phone_dose(tissue, duration_headphones, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_headphones}{duration of time Bluetooth headphones were worn in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from Bluetooth headphone use (phone contribution only) in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from Bluetooth headphone use (phone
contribution only).
}
\details{
The RF-EMF dose from Bluetooth headphone use (phone
contribution) is calculated as:

\deqn{Dose_{phone} = mSAR_{phone} * duration_{phone}}

Where:
\itemize{
\item \eqn{mSAR_{phone}} is the momentary SAR for the headphone connection
\item \eqn{duration_{phone}} is the use duration in s/day
}
}
\examples{
headphones_phone_dose(
  tissue              = "body",
  duration_headphones = 6000)

}
