% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptop.R
\name{laptop_dose}
\alias{laptop_dose}
\title{Calculate Dose from Laptop Use}
\usage{
laptop_dose(
  tissue,
  dur_low,
  dur_lowmed,
  dur_medhigh,
  dur_high,
  params = load_params()
)
}
\arguments{
\item{dur_low}{tba}

\item{dur_lowmed}{tba}

\item{dur_medhigh}{tba}

\item{dur_high}{tba}

\item{params}{Parameter list}
}
\value{
List with brain dose and body dose in mJ/kg/day
}
\description{
Calculate Dose from Laptop Use
}
