% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_params}
\alias{load_params}
\title{Load dose calculation parameters}
\usage{
load_params(path = NULL)
}
\arguments{
\item{path}{Path of parameter file to load (must be a YAML file). If null,
default parameters are loaded.}
}
\value{
List of parameters loaded from yaml input file
}
\description{
Loads default parameters or parameters from external source and returns them
as a nested list.
}
\seealso{
\code{\link[=load_tissue_params]{load_tissue_params()}}
}
