% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_params}
\alias{load_params}
\title{Loading parameters}
\usage{
load_params(path = NULL)
}
\arguments{
\item{path}{Name of parameter file to load (must be in yaml format)}
}
\value{
List of parameters loaded from yaml input file
}
\description{
Loading parameters
}
