% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_tissue_params}
\alias{load_tissue_params}
\title{Load tissue-specific parameters}
\usage{
load_tissue_params(params, device_type, tissue_name)
}
\arguments{
\item{params}{List of parameters (default parameter file or custom YAML that
matches default's structure)}

\item{device_type}{Name of device (default: )}

\item{tissue_name}{Name of tissue (default: "brain" or "body")}
}
\value{
List of parameters (SAR values) specific to selected device type and
tissue.
}
\description{
Loads device- and tissue-specific parameters (SAR values) and returns them as
a list.
}
\seealso{
\code{\link[=load_params]{load_params()}}
}
