% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_dose}
\alias{mobilecall_dose}
\title{Calculate RF-EMF dose from mobile calling (mobile phone and Bluethooth)}
\usage{
mobilecall_dose(
  duration,
  ear_prop,
  headp_prop,
  urbanicity,
  use_5g,
  travel_time,
  headp_ear_num,
  wifi_prop_home,
  wifi_prop_work,
  wifi_prop_travel,
  params = load_params()
)
}
\arguments{
\item{duration}{Duration of mobile phone call in seconds per day}

\item{ear_prop}{Proportion of time mobile phone is held against ear
during call}

\item{headp_prop}{Proportion of time Bluetooth headphones are used during
mobile call}

\item{urbanicity}{Urbanicity of home / workplace}

\item{use_5g}{TRUE if participant uses 5G services on mobile phone,
FALSE if not}

\item{travel_time}{Time spent commuting in seconds per day}

\item{headp_ear_num}{If particpant uses 1 or 2 Bluetooth headphones
during mobile call}

\item{wifi_prop_home}{...}

\item{wifi_prop_work}{...}

\item{wifi_prop_travel}{...}

\item{params}{Parameter list}
}
\value{
List with two values:
\itemize{
\item "brain_call_dose" (brain RF-EMF dose from mobile calls in mJ/kg/day)
\item "body_call_dose" (body RF-EMF dose from mobile calls in mJ/kg/day)
}
}
\description{
Includes all sources: RF-EMF from phone itself, for Bluetooth connection to
headphones, and from headphones themselves
}
\details{
The mobile call RF-EMF dose is calculated as:
\deqn{Dose_{mobilecall} = Dose_{phone} + Dose_{bluetooth}}

Where:
\itemize{
\item \eqn{Dose_{mobilecall}} is the total dose from all mobile call activities
\item \eqn{Dose_{phone}} is the dose from the mobile phone (no Bluetooth)
\item \eqn{Dose_{bluetooth}} is the dose from using Bluetooth headphones
during the call.
}

The dose is calculated as:

\deqn{mSAR * duration}

Where:
\itemize{
\item \eqn{mSAR} is the momentary SAR value in mJ/kg
\item \eqn{duration} is the call duration in seconds
}
}
\seealso{
\code{\link[=mobilecall_msar]{mobilecall_msar()}}
}
