% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_dose}
\alias{mobilecall_dose}
\title{Calculation of RF-EMF Dose from Mobile Phone Calls}
\usage{
mobilecall_dose(
  tissue,
  duration,
  ear_prop,
  headp_prop,
  urbanicity,
  use_5g,
  travel_time,
  headp_ear_num,
  wifi_prop_home,
  wifi_prop_work,
  wifi_prop_travel,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration}{Duration of mobile phone call in seconds per day}

\item{ear_prop}{Proportion of call performed with phone held against ear}

\item{headp_prop}{Proportion of call performed with active Bluetooth connection
to headphones}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{use_5g}{TRUE if 5G services are used for calls, FALSE if not}

\item{travel_time}{Time spent commuting in seconds per day}

\item{headp_ear_num}{Number of Bluetooth headphones used during call}

\item{wifi_prop_home}{Proportion of time connected to WiFi (vs mobile data)
at home}

\item{wifi_prop_work}{Proportion of time connected to WiFi (vs mobile data)
at school / work}

\item{wifi_prop_travel}{Proportion of time connected to WiFi (vs mobile data)
while commuting}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose in mJ/kg/day
}
\description{
Includes RF-EMF exposure from phone itself, for Bluetooth connection to
headphones, and from headphones themselves
}
\details{
The mobile call RF-EMF dose is calculated as:
\deqn{Dose_{mobilecall} = Dose_{phone} + Dose_{bluetooth}}

Where:
\itemize{
\item \eqn{Dose_{mobilecall}} is the total dose from all mobile call activities
\item \eqn{Dose_{phone}} is the dose from the mobile phone (no Bluetooth)
\item \eqn{Dose_{bluetooth}} is the dose from using Bluetooth headphones
during the call.
}

For each source (phone and bluetooth), the dose is calculated as:

\deqn{mSAR * duration}

Where:
\itemize{
\item \eqn{mSAR} is the momentary SAR value in mJ/kg
\item \eqn{duration} is the call duration in seconds
}
}
\examples{
mobilecall_dose(
tissue           = "brain",
duration         = 400,
ear_prop         = 0.67,
headp_prop       = 0.17,
urbanicity       = "suburban",
use_5g           = TRUE,
travel_time      = 1800,
headp_ear_num    = 2,
wifi_prop_home   = 1,
wifi_prop_work   = 0.8,
wifi_prop_travel = 0,
params           = load_params())

}
\seealso{
\code{\link[=mpc_msar]{mpc_msar()}}
}
