% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_msar}
\alias{mobilecall_msar}
\title{Calculate momentary SAR (mSAR) from mobile calling -------------------------}
\usage{
mobilecall_msar(
  tissue,
  prop_native,
  prop_data,
  prop_wifi,
  headp_prop,
  ear_prop,
  speaker_prop,
  urbanicity,
  travel_time,
  params = load_params()
)
}
\description{
Calculate momentary SAR (mSAR) from mobile calling -------------------------
}
