% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mobilecall_msar}
\alias{mobilecall_msar}
\title{Calculate momentary SAR (mSAR) from mobile calling -------------------------}
\usage{
mobilecall_msar(
  tissue,
  prop_native,
  prop_data,
  prop_wifi,
  headp_prop,
  ear_prop,
  speaker_prop,
  urbanicity,
  use_5g,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{tissue}{brain or body}

\item{prop_native}{proportion of native calls}

\item{prop_data}{proportion of data calls}

\item{prop_wifi}{proportion of wifi calls}

\item{headp_prop}{Proportion of time Bluetooth headphones are used during
mobile call}

\item{ear_prop}{Proportion of time mobile phone is held against ear
during call}

\item{speaker_prop}{Proportion of time speaker mode is used during call}

\item{urbanicity}{Urbanicity of home / workplace}

\item{use_5g}{TRUE if participant uses 5G services on mobile phone,
FALSE if not}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list}
}
\description{
Calculate momentary SAR (mSAR) from mobile calling -------------------------
}
