% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{mobiledata_dose}
\alias{mobiledata_dose}
\title{Calculation of RF-EMF Dose from Mobile Phone Data Use}
\usage{
mobiledata_dose(
  tissue,
  duration_low,
  duration_lowmed,
  duration_medhigh,
  duration_high,
  use_5g,
  wifi_prop_home,
  wifi_prop_work,
  wifi_prop_travel,
  urbanicity,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_low}{Duration (in seconds per day) of low output power activities}

\item{duration_lowmed}{Duration (in seconds per day) of low-medium output power activities}

\item{duration_medhigh}{Duration (in seconds per day) of medium-high output power activities}

\item{duration_high}{Duration (in seconds per day) of high output power activities}

\item{use_5g}{TRUE if 5G services are used for data transfer, FALSE if not}

\item{wifi_prop_home}{Proportion of time connected to WiFi (vs mobile data)
at home}

\item{wifi_prop_work}{Proportion of time connected to WiFi (vs mobile data)
at school / work}

\item{wifi_prop_travel}{Proportion of time connected to WiFi (vs mobile data)
while commuting}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from data use in mJ/kg/day
}
\description{
The mobile data RF-EMF dose is calculated as:
\deqn{Dose_{data} = mSAR_{data}*duration_{data}}
}
\details{
Where:
\itemize{
\item \eqn{Dose_{data}} is the dose from mobile data use
\item \eqn{mSAR_{data}} is the momentary SAR value in mJ/kg
\item \eqn{duration_{data}} is the duration of mobile data use
}

We distinguish between 4 types of mobile data use:
\enumerate{
\item Low output power: sending e-mails, browsing the internet, scrolling and
chatting on social media, sending text messages
\item Low to medium output power: online gaming, streaming music, sending voice
messages
\item Medium to high output power: watching videos, uploading pictures or
videos, making video calls
\item High output power: uploading large files
}
}
\examples{
mobiledata_dose(
tissue           = "brain",
duration_low     = 4860,
duration_lowmed  = 540,
duration_medhigh = 4860,
duration_high    = 540,
use_5g           = FALSE,
wifi_prop_home   = 0.5,
wifi_prop_work   = 0.5,
wifi_prop_travel = 0.5,
urbanicity       = "suburban",
travel_time      = 1800)

}
\seealso{
\code{\link[=mpd_msar]{mpd_msar()}}
}
