% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_bt_earset_msar}
\alias{mpc_bt_earset_msar}
\title{Calculate mobile call mSAR for Bluetooth headphones (headphones only)}
\usage{
mpc_bt_earset_msar(tissue, headp_ear_num, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate mSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
mSAR from Bluetooth calls (headphones/earset only) in mJ/kg/day)
}
\description{
Calculates the mSAR for mobile calling using Bluetooth headphones
(headphone/earset only; the contribution of the mobile phone is calculated
separately).
}
\examples{
mpc_bt_earset_msar(
  tissue = "brain"
)

}
\seealso{
\code{\link[=mpc_bt_phone_sar]{mpc_bt_phone_sar()}}
}
