% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_bt_msar}
\alias{mpc_bt_msar}
\title{Calculate call mSAR (Bluetooth contribution from phone and headphones)}
\usage{
mpc_bt_msar(tissue, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate mSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
mSAR from Bluetooth calls in mW/kg
}
\description{
Description TBA
}
\examples{
mpc_bt_msar(
tissue = "brain",
params = load_params()
)

}
