% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_msar}
\alias{mpc_msar}
\title{Calculation of mSAR from Mobile Phone Calls}
\usage{
mpc_msar(
  tissue,
  prop_native,
  prop_data,
  prop_wifi,
  headp_prop,
  ear_prop,
  speaker_prop,
  urbanicity,
  use_5g,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate mSAR (default: "brain" or "body")}

\item{prop_native}{Proportion of mobile call using native connection}

\item{prop_data}{Proportion of mobile call using mobile data connection}

\item{prop_wifi}{Proportion of mobile call using WiFi connection}

\item{headp_prop}{Proportion of call performed with active Bluetooth connection
to headphones}

\item{ear_prop}{Proportion of call performed with phone held against ear}

\item{speaker_prop}{Proportion of call performed in speaker mode}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{use_5g}{TRUE if 5g services are used for calls, FALSE if not}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Momentary SAR (mSAR) in mW/kg
}
\description{
Calculates mSAR (momentary specifc absorption rate) from mobile calls for
a specific tissue.
}
\details{
The mSAR is calculated as

\deqn{mSAR = nSAR_{native}\times outputpower_{native} + nSAR_{data} \times
outputpower_{data} + nSAR_{wifi} \times outputpower_{wifi}}

where
\itemize{
\item \eqn{nSAR_{native}}, \eqn{nSAR_{data}}, \eqn{nSAR_{wifi}} are the nSAR
(normalized specific absorption rates, in W/kg/W) from mobile phone calls using
native, mobile data, or WiFi networks, respectively
\item \eqn{outputpower_{native}}, \eqn{outputpower_{data}},
\eqn{outputpower_{wifi}} are the output powers in mW of the device (mobile
phone) using native, mobile data, or WiFi networks, respectively
}
}
\examples{
mpc_msar(
tissue           = "body",
prop_native      = 0.7,
prop_data        = 0.2,
prop_wifi        = 0.1,
headp_prop       = 0.17,
ear_prop         = 0.67,
speaker_prop     = 0.17,
urbanicity       = "suburban",
use_5g           = TRUE,
travel_time      = 1800)

}
\seealso{
\code{\link[=mpc_pwr_native(), mpc_pwr_data(), mpc_pwr_wifi(), mpc_sar_native(), mpc_sar_data(), mpc_sar_wifi]{mpc_pwr_native(), mpc_pwr_data(), mpc_pwr_wifi(), mpc_sar_native(), mpc_sar_data(), mpc_sar_wifi()}}
}
