% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_pwr_data}
\alias{mpc_pwr_data}
\title{Calculate Mobile Phone Output Power during Data Mobile Phone Calls}
\usage{
mpc_pwr_data(band, urbanicity, travel_time, params = load_params())
}
\arguments{
\item{band}{Technology (2G, 3G, 4G, or 5G) used for the call}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power in mW
}
\description{
Description TBA
}
\details{
The output power depends in the technology used and the location in which the
call is performed (urbanicity, indoors/outdoors/commuting)
}
\examples{
mpc_pwr_data(
band        = "4G",
urbanicity  = "suburban",
travel_time = 1800,
params      = load_params())

}
