% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_pwr_data}
\alias{mpc_pwr_data}
\title{Calculate data call mSAR by technology (2G, 3G, 4G, 5G)}
\usage{
mpc_pwr_data(tech, urbanicity, travel_time, params = load_params())
}
\arguments{
\item{tech}{technology (2G, 3G, 4G, 5G)}

\item{urbanicity}{Urbanicity}

\item{travel_time}{Travel time}

\item{params}{parameter list}
}
\description{
Calculate data call mSAR by technology (2G, 3G, 4G, 5G)
}
