% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_pwr_native}
\alias{mpc_pwr_native}
\title{Calculate native call output power by technology (2G, 3G, 4G, 5G)}
\usage{
mpc_pwr_native(tech, urbanicity, travel_time, params = load_params())
}
\arguments{
\item{tech}{technology (2G, 3G, 4G, 5G)}

\item{urbanicity}{Urbanicity}

\item{travel_time}{Travel time}

\item{params}{parameter list}
}
\description{
Calculate native call output power by technology (2G, 3G, 4G, 5G)
}
