% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_pwr_wifi}
\alias{mpc_pwr_wifi}
\title{Calculate WiFi call output power by band (2.4 GHz, 5.0 GHz)}
\usage{
mpc_pwr_wifi(band, params = load_params())
}
\arguments{
\item{band}{technology (2.4 GHz, 5.0 GHz)}

\item{params}{parameter list}
}
\description{
Calculate WiFi call output power by band (2.4 GHz, 5.0 GHz)
}
