% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_sar_data}
\alias{mpc_sar_data}
\title{Calculate nSAR during Data Mobile Phone Calls}
\usage{
mpc_sar_data(
  tissue,
  band,
  headp_prop,
  ear_prop,
  speaker_prop,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{band}{Technology (2G, 3G, 4G, or 5G) used for the call}

\item{headp_prop}{Proportion of call performed with active Bluetooth connection
to headphones}

\item{ear_prop}{Proportion of call performed with phone held against ear}

\item{speaker_prop}{Proportion of call performed in speaker mode}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
nSAR in W/kg/W
}
\description{
Description TBA
}
\details{
The normalized specific absorption rate (nSAR) depends on the tissue, the
technology (2G, 3G, 4G or 5G), and the location of the phone during the
call (against ear, with Bluetooth headphones, in speaker mode)
}
\examples{
mpc_sar_data(
tissue       = "brain",
band         = "4G",
headp_prop   = 0.17,
ear_prop     = 0.67,
speaker_prop = 0.17,
params       = load_params())

}
