% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilecall.R
\name{mpc_sar_native}
\alias{mpc_sar_native}
\title{Calculate native call SAR by technology (2G, 3G, 4G, 5G)}
\usage{
mpc_sar_native(
  tissue,
  band,
  headp_prop,
  ear_prop,
  speaker_prop,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue}

\item{band}{technology (2G, 3G, 4G, 5G)}

\item{headp_prop}{Headp prop}

\item{ear_prop}{Ear prop}

\item{speaker_prop}{Speaker prop}

\item{params}{parameter list}
}
\description{
Calculate native call SAR by technology (2G, 3G, 4G, 5G)
}
