% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{mpd_pwr_data}
\alias{mpd_pwr_data}
\title{Calculate Mobile Phone Output Power during Mobile Data Use}
\usage{
mpd_pwr_data(
  band,
  duration_low,
  duration_lowmed,
  duration_medhigh,
  duration_high,
  urbanicity,
  travel_time,
  params = load_params()
)
}
\arguments{
\item{band}{Technology (3g, 4g, or 5g)}

\item{duration_low}{Duration (in seconds per day) of low output power activities}

\item{duration_lowmed}{Duration (in seconds per day) of low-medium output power activities}

\item{duration_medhigh}{Duration (in seconds per day) of medium-high output power activities}

\item{duration_high}{Duration (in seconds per day) of high output power activities}

\item{urbanicity}{Urbanicity of home / workplace (rural, suburban, or urban)}

\item{travel_time}{Time spent commuting in seconds per day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Output power in mW
}
\description{
Calculates the mobile phone output power during mobile data use.
}
\details{
The output power depends in the technology used, the location in which the
mobile data network (urbanicity, indoors/outdoors/commuting), and the activity
the phone is used for.
}
\examples{
mpd_pwr_data(
band             = "4g",
duration_low     = 4860,
duration_lowmed  = 540,
duration_medhigh = 4860,
duration_high    = 540,
urbanicity       = "rural",
travel_time      = 1800)

}
