% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledata.R
\name{mpd_sar_wifi}
\alias{mpd_sar_wifi}
\title{Calculate nSAR during WiFi Use on mobile phone}
\usage{
mpd_sar_wifi(band, tissue, params = load_params())
}
\arguments{
\item{band}{Frequency band ("2" for 2.4 GHz, "5" for 5.0 GHz)}

\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
nSAR in W/kg/W
}
\description{
Calculates tissue-specific nSAR from WiFi use on mobile phone.
}
\details{
The normalized specific absorption rate (nSAR) depends on the tissue and
the frequency band.
}
\examples{
mpd_sar_wifi(
tissue       = "brain",
band         = "2",
params       = load_params())

}
