% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{other_dose_wrapper}
\alias{other_dose_wrapper}
\title{Calculate RF-EMF Dose from Other Sources}
\usage{
other_dose_wrapper(
  tissue,
  duration_hotspot,
  duration_smartwatch,
  duration_vr,
  duration_headphones,
  duration_gaming,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{duration_hotspot}{Duration of time hotspot was active in s/day}

\item{duration_smartwatch}{Duration of time smartwatch was used in s/day}

\item{duration_vr}{Duration of time VR headset was used in s/day}

\item{duration_headphones}{Duration of time Bluetooth headphones were used in s/day}

\item{duration_gaming}{Duration of gaming with portable console in s/day}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from other sources in mJ/kg/day
}
\description{
Wrapper function that calculates total dose from additional devices and
activities, as used in the GOLIAT consortium study dose calculation.
}
\details{
The RF-EMF dose from other sources takes into account 5 additional sources:
\enumerate{
\item Hotspot (smartphone)
\item Smart watch
\item Virtual reality (VR) headset
\item Bluetooth headphones
\item Portable gaming consoles
}

The dose is calculated as

\deqn{Dose_{other} = Dose_{hotspot} + Dose_{smartwatch} + Dose_{vr} + Dose_{headphones} + Dose_{gaming}}

The dose from each source is calculated from the corresponding dose functions.
}
\examples{
other_dose_wrapper(
  tissue              = "brain",
  duration_hotspot    = 600,
  duration_smartwatch = 86400,
  duration_vr         = 0,
  duration_headphones = 8280,
  duration_gaming     = 3600)

}
\seealso{
\code{\link[=hotspot_dose(), smartwatch_dose(), vr_dose(), headphones_dose(), gaming_dose]{hotspot_dose(), smartwatch_dose(), vr_dose(), headphones_dose(), gaming_dose()}}
}
