% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablet.R
\name{tablet_dose}
\alias{tablet_dose}
\title{Calculate RF-EMF Dose from Tablet Use}
\usage{
tablet_dose(
  tissue,
  dur_low,
  dur_lowmed,
  dur_medhigh,
  dur_high,
  params = load_params()
)
}
\arguments{
\item{tissue}{Tissue for which to calculate dose (default: "brain" or "body")}

\item{dur_low}{Duration (in seconds per day) of low output power activities on tablet}

\item{dur_lowmed}{Duration (in seconds per day) of low-medium output power activities on tablet}

\item{dur_medhigh}{Duration (in seconds per day) of medium-high output power activities on tablet}

\item{dur_high}{Duration (in seconds per day) of high output power activities on tablet}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
Tissue-specific RF-EMF dose from tablet use in mJ/kg/day
}
\description{
Calculates the tissue-specific RF-EMF dose from tablet use.
}
\details{
The tablet RF-EMF dose is calculated as:

\deqn{Dose_{tablet} = mSAR_{tablet}*duration_{tablet}}

Where:
\itemize{
\item \eqn{Dose_{tablet}} is the dose from tablet use
\item \eqn{mSAR_{tablet}} is the momentary SAR value in mJ/kg
\item \eqn{duration_{tablet}} is the duration of tablet use
}

We distinguish between 4 types of activities:
\enumerate{
\item Low output power: sending e-mails, browsing the internet, scrolling and
chatting on social media, sending text messages
\item Low to medium output power: online gaming, streaming music, sending voice
messages
\item Medium to high output power: watching videos, uploading pictures or
videos, making video calls
\item High output power: uploading large files
}
}
\examples{
tablet_dose(
  tissue      = "brain",
  dur_low     = 0,
  dur_lowmed  = 681,
  dur_medhigh = 0,
  dur_high    = 0)

}
\seealso{
\code{\link[=tablet_msar]{tablet_msar()}}
}
