% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\name{wifi_msar}
\alias{wifi_msar}
\title{Calculate mSAR of WiFi Router}
\usage{
wifi_msar(tissue, travel_time, wifi_prop_travel, params = load_params())
}
\arguments{
\item{tissue}{Tissue for which to calculate nSAR (default: "brain" or "body")}

\item{travel_time}{Time spent commuting in seconds per day}

\item{wifi_prop_travel}{Proportion of time connected to WiFi (vs mobile data)
while commuting}

\item{params}{Parameter list (optional). If not specified, calculations use
default parameters.}
}
\value{
mSAR in W/kg/W/m**2
}
\description{
Calculates the mSAR of WiFi Router
}
\details{
The output power depends on the time spent commuting and the proportion of
time a WiFi connection (vs mobile data) is used while commuting.
}
\examples{
wifi_msar(
  tissue           = "body",
  travel_time      = 1800,
  wifi_prop_travel = 0.5)

}
\seealso{
\code{\link[=wifi_pwr(), wifi_sar]{wifi_pwr(), wifi_sar()}}
}
